% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{AD-class}
\alias{AD-class}
\alias{.AD}
\title{AD}
\description{
An S4 class to represent the (Gregorian) AD era.
}
\seealso{
Other classes: 
\code{\link{BC-class}},
\code{\link{BCE-class}},
\code{\link{BP-class}},
\code{\link{CE-class}},
\code{\link{GregorianCalendar-class}},
\code{\link{JulianCalendar-class}},
\code{\link{RataDie-class}},
\code{\link{TimeScale-class}},
\code{\link{TimeSeries-class}},
\code{\link{b2k-class}}

Other gregorian era classes: 
\code{\link{BC-class}},
\code{\link{BCE-class}},
\code{\link{BP-class}},
\code{\link{CE-class}},
\code{\link{GregorianCalendar-class}},
\code{\link{b2k-class}}
}
\author{
N. Frerebeau
}
\concept{classes}
\concept{gregorian era classes}
\keyword{internal}
