% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aion-package.R
\docType{package}
\name{aion-package}
\alias{aion}
\alias{aion-package}
\title{aion: Cyclical and recurring pattern analysis}
\description{
Tools that aid in the evaluation of recurring patterns particularly with recurrent events. This is both from a data tidying and analytical perspective.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/asshah4/aion}
  \item Report bugs at \url{https://github.com/asshah4/aion/issues}
}

}
\author{
\strong{Maintainer}: Anish Shah \email{ashah282@uic.edu} (\href{https://orcid.org/0000-0002-9729-1558}{ORCID})

}
\keyword{internal}
