% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/allen.R
\docType{methods}
\name{allen_intersect}
\alias{allen_intersect}
\alias{allen_intersect-method}
\alias{allen_intersect,character,character-method}
\title{Intersection of Allen Relations}
\usage{
allen_intersect(x, y, ...)

\S4method{allen_intersect}{character,character}(x, y)
}
\arguments{
\item{x, y}{A \code{\link{character}} vector of Allen relations.}

\item{...}{Currently not used.}
}
\value{
A \code{\link{character}} vector.
}
\description{
Intersection of Allen Relations
}
\examples{
## Complement
(comp <- allen_complement("F"))

## Converse
(conv <- allen_converse(comp))

## Composition
allen_composition("oFD", "oFDseS")

## Intersection
allen_intersect("pFsSf", "pmoFD")

# Union
allen_union("pFsSf", "pmoFD")
}
\references{
Allen, J. F. (1983). Maintaining Knowledge about Temporal Intervals.
\emph{Communications of the ACM}, 26(11): 832-843. \doi{10.1145/182.358434}.
}
\seealso{
Other Allen's intervals: 
\code{\link{allen_complement}()},
\code{\link{allen_composition}()},
\code{\link{allen_converse}()},
\code{\link{allen_relation}()},
\code{\link{allen_relation_code}()},
\code{\link{allen_union}()}
}
\author{
T. S. Dye, N. Frerebeau
}
\concept{Allen's intervals}
