% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/calendar.R
\docType{methods}
\name{as_calendar}
\alias{as_calendar}
\alias{as_gregorian}
\alias{as_julian}
\title{Create a Calendar}
\usage{
as_gregorian(label, name, epoch, direction)

as_julian(...)
}
\arguments{
\item{label}{A \code{\link{character}} string specifying the abbreviated label of
the time scale.}

\item{name}{A \code{\link{character}} string specifying the name of the time scale.}

\item{epoch}{A length-one \code{\link{numeric}} vector specifying the epoch year from
which years are counted (in Gregorian astronomical years).}

\item{direction}{A length-one \code{\link{integer}} vector specifying if years are
counted backwards (\code{-1}) or forwards (\code{1}) from \code{epoch}. Only the
\link[=sign]{sign} of \code{direction} will be retained.}

\item{...}{Currently not used.}
}
\value{
A \code{\linkS4class{Calendar}} object.
}
\description{
Create a Calendar
}
\note{
Inspired by \code{\link[era:era]{era::era()}} by Joe Roe.
}
\examples{
## Define time scales
(BP <- calendar("BP"))
(AD <- calendar("AD"))

## Create a custom gregorian calendar
as_gregorian(
  label = "cal BP",
  name = "Before Present",
  epoch = 1950,
  direction = -1
)
}
\seealso{
Other time scales tools: 
\code{\link{calendar_get}},
\code{\link{calendar}()},
\code{\link{convert}()}
}
\author{
N. Frerebeau
}
\concept{time scales tools}
