% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/years.R
\docType{methods}
\name{as_date}
\alias{as_date}
\alias{as_date-method}
\alias{as_date,numeric,GregorianCalendar-method}
\title{Date Conversion from \emph{Rata Die}}
\usage{
as_date(object, calendar)

\S4method{as_date}{numeric,GregorianCalendar}(object, calendar)
}
\arguments{
\item{object}{A \code{\linkS4class{RataDie}} object (see \code{\link[=as_fixed]{as_fixed()}}).}

\item{calendar}{A \code{\linkS4class{TimeScale}} object (see \code{\link[=calendar]{calendar()}}).}
}
\description{
Date Conversion from \emph{Rata Die}
}
\examples{
## R foundation
(y <- as_fixed(year = 1994, calendar = calendar("CE")))
as_year(y, calendar = calendar("CE"))

## Vector of years expressed in ka BP
(ka <- as_fixed(c(30, 35, 40), calendar = calendar("BP"), scale = 1000))
as_year(ka, calendar = calendar("CE"))
}
\references{
Reingold, E. M. and Dershowitz, N. (2018). \emph{Calendrical Calculations:
The Ultimate Edition}. Cambridge University Press.
\doi{10.1017/9781107415058}.
}
\seealso{
Other fixed date tools: 
\code{\link{as_fixed}()},
\code{\link{as_year}()}
}
\author{
N. Frerebeau
}
\concept{fixed date tools}
