% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/calendar-gregorian.R
\docType{methods}
\name{as_fixed}
\alias{as_fixed}
\alias{as_fixed-method}
\alias{as_fixed,numeric,missing,missing,GregorianCalendar-method}
\alias{as_fixed,numeric,numeric,numeric,GregorianCalendar-method}
\title{\emph{Rata Die}}
\usage{
as_fixed(year, month, day, calendar, ...)

\S4method{as_fixed}{numeric,missing,missing,GregorianCalendar}(year, calendar, scale = 1)

\S4method{as_fixed}{numeric,numeric,numeric,GregorianCalendar}(year, month, day, calendar)
}
\arguments{
\item{year}{A \code{\link{numeric}} vector of (decimal) years.}

\item{month}{A \code{\link{numeric}} vector of (decimal) years.}

\item{day}{A \code{\link{numeric}} vector of (decimal) years.}

\item{calendar}{A \code{\linkS4class{TimeScale}} object (see \code{\link[=calendar]{calendar()}}).}

\item{...}{Currently not used.}

\item{scale}{A length-one \code{\link{integer}} vector specifying the number of years
represented by one unit. It should be a power of 10 (i.e. 1000 means ka).}
}
\value{
A \code{\linkS4class{RataDie}} object.
}
\description{
\emph{Rata Die}
}
\details{
\emph{Rata die} are represented as the number of days since 01-01-01 (Gregorian),
with negative values for earlier dates.
}
\examples{
## R 1.0.0
(y <- as_fixed(year = 2000, month = 02, day = 29, calendar = CE()))
as_date(y, calendar = CE())
as_year(y, calendar = CE())

## Vector of years expressed in ka BP
(ka <- as_fixed(c(30, 35, 40), calendar = BP(), scale = 1000))
as_year(ka, calendar = CE())
}
\references{
Reingold, E. M. and Dershowitz, N. (2018). \emph{Calendrical Calculations:
The Ultimate Edition}. Cambridge University Press.
\doi{10.1017/9781107415058}.
}
\seealso{
Other fixed date tools: 
\code{\link{as_date}()},
\code{\link{as_year}()}
}
\author{
N. Frerebeau
}
\concept{fixed date tools}
