% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plot.R
\docType{methods}
\name{axis_year}
\alias{axis_year}
\alias{axis_year-method}
\alias{axis_year,RataDie-method}
\alias{axis_year,TimeSeries-method}
\title{Time Series Plotting Functions}
\usage{
axis_year(x, ...)

\S4method{axis_year}{RataDie}(
  x,
  side,
  at = NULL,
  format = c("a", "ka", "Ma", "Ga"),
  labels = TRUE,
  calendar = getOption("chronos.calendar"),
  ...
)

\S4method{axis_year}{TimeSeries}(
  x,
  side,
  at = NULL,
  format = c("a", "ka", "Ma", "Ga"),
  labels = TRUE,
  calendar = getOption("chronos.calendar"),
  ...
)
}
\arguments{
\item{x}{A \code{\linkS4class{TimeSeries}} object.}

\item{...}{Further parameters to be passed to \code{\link[graphics:axis]{graphics::axis()}}.
(e.g. \link[graphics:par]{graphical parameters}).}

\item{side}{An \code{\link{integer}} specifying which side of the plot the axis is to
be drawn on. The axis is placed as follows: 1=below, 2=left, 3=above and
4=right.}

\item{at}{A \code{\link{numeric}} vector giving the points at which tick-marks are to
be drawn. If \code{NULL}, tickmark locations are computed.}

\item{format}{A \code{\link{character}} string specifying the prefix.
It should be one of "\code{a}", "\code{ka}", "\code{Ma}" or "\code{Ga}".
If \code{TRUE}, a good guess for an appropriate format is made.}

\item{labels}{A \code{\link{logical}} scalar specifying whether annotations are to be
made at the tickmarks, or a vector of \code{\link{character}} strings to be placed at
the tickpoints.}

\item{calendar}{A \code{\linkS4class{TimeScale}} object specifying the target calendar
(see \code{\link[=calendar]{calendar()}}).}
}
\value{
\code{axis_year()} is called it for its side-effects. Invisibly returns \code{x}.
}
\description{
Time Series Plotting Functions
}
\examples{
## Create 6 time-series of 50 observations
## Sampled every two years starting from 2000 BP
X <- series(
  object = matrix(rnorm(300), nrow = 50, ncol = 6),
  time = seq(2000, by = 2, length.out = 50),
  calendar = BP()
)

## Multiple
plot(X) # Default calendar
plot(X, calendar = BP(), flip = TRUE) # BP
plot(X, calendar = b2k(), ncol = 1) # b2k

## Single
plot(X, type = "single", calendar = BP(), col = rainbow(6)) # BP
}
\seealso{
Other time series tools: 
\code{\link{data.frame}},
\code{\link{plot}()},
\code{\link{series}()},
\code{\link{span}()},
\code{\link{start}()},
\code{\link{time}()},
\code{\link{window}()}
}
\author{
N. Frerebeau
}
\concept{time series tools}
