% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plot.R
\docType{methods}
\name{c14_plot}
\alias{c14_plot}
\alias{plot,CalibratedAges,missing-method}
\alias{plot,CalibratedSPD,missing-method}
\title{Plot Calibrated Radiocarbon Ages}
\usage{
\S4method{plot}{CalibratedAges,missing}(
  x,
  density = TRUE,
  interval = TRUE,
  level = 0.954,
  warnings = TRUE,
  decreasing = TRUE,
  main = NULL,
  sub = NULL,
  ann = graphics::par("ann"),
  axes = TRUE,
  frame.plot = FALSE,
  panel.first = NULL,
  panel.last = NULL,
  ...
)

\S4method{plot}{CalibratedSPD,missing}(
  x,
  xlim = NULL,
  ylim = NULL,
  main = NULL,
  sub = NULL,
  ann = graphics::par("ann"),
  axes = TRUE,
  frame.plot = FALSE,
  panel.first = NULL,
  panel.last = NULL,
  ...
)
}
\arguments{
\item{x}{A \code{\linkS4class{CalibratedAges}} or \code{\linkS4class{CalibratedSPD}} object.}

\item{density}{A \code{\link{logical}} scalar: should density be drawn?}

\item{interval}{A \code{\link{logical}} scalar: should highest posterior density
interval be drawn?}

\item{level}{A length-one \code{\link{numeric}} vector giving the confidence level.}

\item{warnings}{A \code{\link{logical}} scalar: should warnings be plotted?}

\item{decreasing}{A \code{\link{logical}} scalar: should the sort order be decreasing?}

\item{main}{A \code{\link{character}} string giving a main title for the plot.}

\item{sub}{A \code{\link{character}} string giving a subtitle for the plot.}

\item{ann}{A \code{\link{logical}} scalar: should the default annotation (title and x,
y and z axis labels) appear on the plot?}

\item{axes}{A \code{\link{logical}} scalar: should axes be drawn on the plot?}

\item{frame.plot}{A \code{\link{logical}} scalar: should a box be drawn around the
plot?}

\item{panel.first}{An an \code{expression} to be evaluated after the plot axes are
set up but before any plotting takes place. This can be useful for drawing
background grids.}

\item{panel.last}{An \code{expression} to be evaluated after plotting has taken
place but before the axes, title and box are added.}

\item{...}{Other \link[graphics:par]{graphical parameters} may also be passed as
arguments to this function.}

\item{xlim}{A length-two \code{\link{numeric}} vector giving the x limits of the plot.}

\item{ylim}{A length-two \code{\link{numeric}} vector giving the y limits of the plot.}
}
\value{
\code{plot()} is called it for its side-effects: it results in a graphic
being displayed. Invisibly returns \code{x}.
}
\description{
Plot Calibrated Radiocarbon Ages
}
\examples{
## Calibrate a single date
cal <- c14_calibrate(300, 20)
plot(cal)

## Calibrate multiple dates
cal <- c14_calibrate(
  ages = c(5000, 4500),
  errors = c(45, 35),
  names = c("X", "Y")
)
plot(cal, panel.first = graphics::grid())

plot(cal, density = TRUE, interval = FALSE)
plot(cal, density = FALSE, interval = TRUE)

## Convert BP scale to CE
CE <- project(cal, "CE")
plot(CE)

## Convert BP scale to b2k
b2k <- project(cal, "b2k")
plot(b2k)
}
\seealso{
Other radiocarbon tools: 
\code{\link{F14C}},
\code{\link{c14_calibrate}()},
\code{\link{c14_combine}()},
\code{\link{c14_curve}()},
\code{\link{spd}()}
}
\author{
N. Frerebeau
}
\concept{radiocarbon tools}
