% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Calendar-class}
\alias{Calendar-class}
\alias{.Calendar}
\title{Calendar}
\description{
An S4 class to represent a year unit.
}
\section{Slots}{

\describe{
\item{\code{unit}}{A \code{\link{character}} string giving the name of the year unit.}

\item{\code{days}}{A \code{\link{numeric}} vector giving the average length of the year in
solar days.}
}}

\seealso{
Other classes: 
\code{\link{CalibratedAges-class}},
\code{\link{CalibratedSPD-class}},
\code{\link{TimeScale-class}}
}
\author{
N. Frerebeau
}
\concept{classes}
\keyword{internal}
