% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/calendar.R
\docType{methods}
\name{calendar}
\alias{calendar}
\alias{calendar-method}
\alias{calendar,character-method}
\alias{calendar,TimeLine-method}
\alias{calendar,TimeSeries-method}
\title{Calendar}
\usage{
calendar(object, ...)

\S4method{calendar}{character}(object)

\S4method{calendar}{TimeLine}(object)

\S4method{calendar}{TimeSeries}(object)
}
\arguments{
\item{object}{A \code{\link{character}} string specifying the time scale (see details).}

\item{...}{Currently not used.}
}
\value{
A \code{\linkS4class{Calendar}} object.
}
\description{
Calendar
}
\details{
The following time scales are available:
\describe{
\item{\code{BP}}{Before Present.}
\item{\code{BC}}{Before Christ.}
\item{\code{BCE}}{Before Common Era.}
\item{\code{AD}}{Anno Domini.}
\item{\code{CE}}{Common Era.}
\item{\code{b2k}}{Years before 2000.}
}
}
\note{
Inspired by \code{\link[era:era]{era::era()}} by Joe Roe.
}
\examples{
## Define time scales
(BP <- calendar("BP"))
(AD <- calendar("AD"))

## Create a custom gregorian calendar
as_gregorian(
  label = "cal BP",
  name = "Before Present",
  epoch = 1950,
  direction = -1
)
}
\seealso{
Other time scales tools: 
\code{\link{as_calendar}},
\code{\link{calendar_get}},
\code{\link{convert}()}
}
\author{
N. Frerebeau
}
\concept{time scales tools}
