% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/calendar.R
\docType{methods}
\name{calendar_get}
\alias{calendar_get}
\alias{calendar_label}
\alias{calendar_label-method}
\alias{calendar_name}
\alias{calendar_name-method}
\alias{calendar_epoch}
\alias{calendar_epoch-method}
\alias{calendar_direction}
\alias{calendar_direction-method}
\alias{calendar_year}
\alias{calendar_year-method}
\alias{calendar_label,Calendar-method}
\alias{calendar_label,TimeLine-method}
\alias{calendar_label,TimeSeries-method}
\alias{calendar_name,Calendar-method}
\alias{calendar_name,TimeLine-method}
\alias{calendar_name,TimeSeries-method}
\alias{calendar_epoch,Calendar-method}
\alias{calendar_epoch,TimeLine-method}
\alias{calendar_epoch,TimeSeries-method}
\alias{calendar_direction,Calendar-method}
\alias{calendar_direction,TimeLine-method}
\alias{calendar_direction,TimeSeries-method}
\alias{calendar_year,Calendar-method}
\alias{calendar_year,TimeLine-method}
\alias{calendar_year,TimeSeries-method}
\title{Calendar Tools}
\usage{
calendar_label(object)

calendar_name(object)

calendar_epoch(object)

calendar_direction(object)

calendar_year(object)

\S4method{calendar_label}{Calendar}(object)

\S4method{calendar_label}{TimeLine}(object)

\S4method{calendar_label}{TimeSeries}(object)

\S4method{calendar_name}{Calendar}(object)

\S4method{calendar_name}{TimeLine}(object)

\S4method{calendar_name}{TimeSeries}(object)

\S4method{calendar_epoch}{Calendar}(object)

\S4method{calendar_epoch}{TimeLine}(object)

\S4method{calendar_epoch}{TimeSeries}(object)

\S4method{calendar_direction}{Calendar}(object)

\S4method{calendar_direction}{TimeLine}(object)

\S4method{calendar_direction}{TimeSeries}(object)

\S4method{calendar_year}{Calendar}(object)

\S4method{calendar_year}{TimeLine}(object)

\S4method{calendar_year}{TimeSeries}(object)
}
\arguments{
\item{object}{A \code{\linkS4class{Calendar}} object.}
}
\description{
Calendar Tools
}
\examples{
## Define time scales
(BP <- calendar("BP"))
(AD <- calendar("AD"))

## Create a custom gregorian calendar
as_gregorian(
  label = "cal BP",
  name = "Before Present",
  epoch = 1950,
  direction = -1
)
}
\seealso{
Other time scales tools: 
\code{\link{as_calendar}},
\code{\link{calendar}()},
\code{\link{convert}()}
}
\author{
N. Frerebeau
}
\concept{time scales tools}
