% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/operators.R
\docType{methods}
\name{compare}
\alias{compare}
\alias{Compare,TimeLine,TimeLine-method}
\title{Chronological Comparison}
\usage{
\S4method{Compare}{TimeLine,TimeLine}(e1, e2)
}
\arguments{
\item{e1, e2}{A \code{\linkS4class{TimeLine}} object.}
}
\value{
A \code{\link{logical}} vector.
}
\description{
Operators performing chronological comparison (not numerical).
}
\details{
The following operators allow to compare two objects. Remember that this is
a \strong{chronological comparison} and not a numerical comparison (e.g. 5000 is
numerically greater than 3000, but 5000 BP precedes 3000 BP).

\tabular{ll}{
\strong{operator} \tab \strong{meaning} \cr
\code{==}         \tab equals \cr
\code{!=}         \tab differs \cr
\code{>}          \tab preceded by \cr
\code{<}          \tab precedes \cr
\code{>=}         \tab preceded by or equals \cr
\code{<=}         \tab precedes or equals \cr
}
}
\examples{
## /!\ Be carefull: chronological comparison (not numerical) /!\

## Vectors of years
(x_CE <- years(c(-9551, -7917, -8387, -3371, -3004), era("CE")))
(y_CE <- years(c(-3940, -4055, -8528, -5068, -8442), era("CE")))

## x precedes y?
x_CE < y_CE

## x preceded by y?
x_CE > y_CE

## Change the origin
(x_BP <- project(x_CE, era("BP")))
(y_BP <- project(y_CE, era("BP")))

## x precedes y?
x_BP < y_BP

all.equal(x_CE < y_CE, x_BP < y_BP)
}
\seealso{
Other operators: 
\code{\link{arithmetic}}
}
\author{
N. Frerebeau
}
\concept{operators}
