% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dates}
\alias{dates}
\title{Sample Data from Reingold and Dershowitz (2018)}
\format{
A \code{\link{data.frame}} with 33 rows and 14 variables:
\describe{
\item{\code{rata_die}}{Rata die.}
\item{\code{weekday}}{Week day.}
\item{\code{jd}}{Julian day.}
\item{\code{mjd}}{Modified Julian day.}
\item{\code{unix}}{Unix.}
\item{\code{gregorian_year}, \code{gregorian_month}, \code{gregorian_day}}{Gregorian date.}
\item{\code{julian_year}, \code{julian_month}, \code{julian_day}}{Julian date.}
\item{\code{egyptian_year}, \code{egyptian_month}, \code{egyptian_day}}{Egyptian date.}
}
}
\usage{
dates
}
\description{
A dataset of 33 dates from the years -1000 to 2100 with their equivalents on
differents calendars.
}
\references{
Reingold, E. M. and Dershowitz, N. (2018). \emph{Calendrical Calculations:
The Ultimate Edition}. Cambridge University Press.
\doi{10.1017/9781107415058}.
}
\concept{datasets}
\keyword{datasets}
