% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/era.R
\docType{methods}
\name{era}
\alias{era}
\alias{era-method}
\alias{era,character-method}
\alias{era,TimeLine-method}
\alias{era,TimeSeries-method}
\title{Era}
\usage{
era(object, ...)

\S4method{era}{character}(object)

\S4method{era}{TimeLine}(object)

\S4method{era}{TimeSeries}(object)
}
\arguments{
\item{object}{A \code{\link{character}} string specifying the time scale (see details).}

\item{...}{Currently not used.}
}
\value{
A \code{\linkS4class{TimeScale}} object.
}
\description{
Era
}
\details{
The following time scales are available:

\tabular{lll}{
\strong{label} \tab \strong{name}           \tab \strong{calendar} \cr
\code{BP}      \tab Before Present     \tab Gregorian    \cr
\code{BC}      \tab Before Christ      \tab Gregorian    \cr
\code{BCE}     \tab Before Common Era  \tab Gregorian    \cr
\code{AD}      \tab Anno Domini        \tab Gregorian    \cr
\code{CE}      \tab Common Era         \tab Gregorian    \cr
\code{b2k}     \tab Years before 2000  \tab Gregorian    \cr
}
}
\note{
Inspired by \code{\link[era:era]{era::era()}} by Joe Roe.
}
\examples{
## Define time scales
(BP <- era("BP"))
(AD <- era("AD"))
}
\seealso{
Other time scales tools: 
\code{\link{convert}()},
\code{\link{era_get}},
\code{\link{gregorian}},
\code{\link{julian}()}
}
\author{
N. Frerebeau
}
\concept{time scales tools}
