% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/era.R
\docType{methods}
\name{era_get}
\alias{era_get}
\alias{era_label}
\alias{era_label-method}
\alias{era_name}
\alias{era_name-method}
\alias{era_epoch}
\alias{era_epoch-method}
\alias{era_scale}
\alias{era_scale-method}
\alias{era_direction}
\alias{era_direction-method}
\alias{era_unit}
\alias{era_unit-method}
\alias{era_days}
\alias{era_days-method}
\alias{era_label,TimeScale-method}
\alias{era_name,TimeScale-method}
\alias{era_epoch,TimeScale-method}
\alias{era_scale,TimeScale-method}
\alias{era_direction,TimeScale-method}
\alias{era_unit,TimeScale-method}
\alias{era_days,TimeScale-method}
\title{Time Scales Mutators}
\usage{
era_label(object)

era_name(object)

era_epoch(object)

era_scale(object)

era_direction(object)

era_unit(object)

era_days(object)

\S4method{era_label}{TimeScale}(object)

\S4method{era_name}{TimeScale}(object)

\S4method{era_epoch}{TimeScale}(object)

\S4method{era_scale}{TimeScale}(object)

\S4method{era_direction}{TimeScale}(object)

\S4method{era_unit}{TimeScale}(object)

\S4method{era_days}{TimeScale}(object)
}
\arguments{
\item{object}{A \code{\linkS4class{TimeScale}} object.}
}
\description{
Time Scales Mutators
}
\note{
Inspired by \code{\link[era:era]{era::era()}} by Joe Roe.
}
\examples{
## Define time scales
(BP <- era("BP"))
(AD <- era("AD"))
}
\seealso{
Other time scales tools: 
\code{\link{convert}()},
\code{\link{era}()}
}
\author{
N. Frerebeau
}
\concept{time scales tools}
