% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/radiocarbon.R
\docType{methods}
\name{F14C}
\alias{F14C}
\alias{BP14C_to_F14C}
\alias{BP14C_to_F14C-method}
\alias{F14C_to_BP14C}
\alias{F14C_to_BP14C-method}
\alias{BP14C_to_F14C,numeric,numeric-method}
\alias{F14C_to_BP14C,numeric,numeric-method}
\title{F14C}
\usage{
BP14C_to_F14C(ages, errors, ...)

F14C_to_BP14C(ratios, errors, ...)

\S4method{BP14C_to_F14C}{numeric,numeric}(ages, errors)

\S4method{F14C_to_BP14C}{numeric,numeric}(ratios, errors)
}
\arguments{
\item{ages}{A \code{\link{numeric}} vector giving the radiocarbon ages.}

\item{errors}{A \code{\link{numeric}} vector giving the standard deviation of the
ages/ratios.}

\item{...}{Currently not used.}

\item{ratios}{A \code{\link{numeric}} vector giving the F14C ratios.}
}
\value{
A two-column \code{\link{data.frame}}.
}
\description{
Converts F14C ratio to 14C age.
}
\references{
Bronk Ramsey, C. (2008). RADIOCARBON DATING: REVOLUTIONS IN UNDERSTANDING.
\emph{Archaeometry}, 50:249-275. \doi{10.1111/j.1475-4754.2008.00394.x}.
}
\seealso{
Other radiocarbon tools: 
\code{\link{c14_calibrate}()},
\code{\link{c14_combine}()},
\code{\link{c14_curve}()},
\code{\link{c14_plot}},
\code{\link{spd}()}
}
\author{
N. Frerebeau
}
\concept{radiocarbon tools}
