% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/calendar-gregorian.R
\docType{methods}
\name{fixed_gregorian}
\alias{fixed_gregorian}
\alias{fixed_from_BP}
\alias{fixed_to_BP}
\alias{fixed_from_BC}
\alias{fixed_to_BC}
\alias{fixed_from_BCE}
\alias{fixed_to_BCE}
\alias{fixed_from_AD}
\alias{fixed_to_AD}
\alias{fixed_from_CE}
\alias{fixed_to_CE}
\alias{fixed_from_b2k}
\alias{fixed_to_b2k}
\title{Gregorian Year Conversion to and from \emph{Rata Die}}
\usage{
fixed_from_BP(year, month, day)

fixed_to_BP(object)

fixed_from_BC(year, month, day)

fixed_to_BC(object)

fixed_from_BCE(year, month, day)

fixed_to_BCE(object)

fixed_from_AD(year, month, day)

fixed_to_AD(object)

fixed_from_CE(year, month, day)

fixed_to_CE(object)

fixed_from_b2k(year, month, day)

fixed_to_b2k(object)
}
\arguments{
\item{year}{A \code{\link{numeric}} vector of years. If \code{month} and \code{day} are missing,
decimal years are expected.}

\item{month}{A \code{\link{numeric}} vector of months.}

\item{day}{A \code{\link{numeric}} vector of days.}

\item{object}{A \code{\linkS4class{RataDie}} object (see \code{\link[=fixed]{fixed()}}).}
}
\value{
\itemize{
\item \verb{fixed_from_*()} returns a \code{\linkS4class{RataDie}} object.
\item \verb{fixed_to_*()} returns
}
}
\description{
Convenient functions for conversion from and to \emph{rata die} for a given
Gregorian era.
}
\examples{
## R 1.0.0
(y <- fixed(year = 2000, month = 02, day = 29, calendar = CE()))
as_date(y, calendar = CE())
as_year(y, calendar = CE())

## Vector of years expressed in ka BP
(ka <- fixed(c(30, 35, 40), calendar = BP(), scale = 1000))
as_year(ka, calendar = CE())
}
\references{
Reingold, E. M. and Dershowitz, N. (2018). \emph{Calendrical Calculations:
The Ultimate Edition}. Cambridge University Press.
\doi{10.1017/9781107415058}.
}
\seealso{
Other fixed date tools: 
\code{\link{as_date}()},
\code{\link{as_decimal}()},
\code{\link{as_year}()},
\code{\link{fixed}()}
}
\author{
N. Frerebeau
}
\concept{fixed date tools}
