% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/graph.R
\docType{methods}
\name{graph}
\alias{graph}
\alias{graph-method}
\alias{graph,matrix-method}
\alias{graph,TimeIntervals-method}
\title{Graph}
\usage{
graph(x, ...)

\S4method{graph}{matrix}(
  x,
  type = c("interval", "stratigraphy"),
  direction = c("above", "below"),
  simplify = TRUE,
  reduce = TRUE,
  ...
)

\S4method{graph}{TimeIntervals}(
  x,
  type = c("interval", "stratigraphy"),
  simplify = TRUE,
  reduce = TRUE,
  ...
)
}
\arguments{
\item{x}{A \code{\linkS4class{TimeIntervals}} object or a two-columns \code{character}
\code{\link{matrix}} where each row specifies one relation element.}

\item{...}{Currently not used.}

\item{type}{A \code{\link{character}} string specifying the type of the graph to be
computed. It must be one of "\code{interval}" (the default) or "\code{stratigraphy}"
(see details). Any unambiguous substring can be given.}

\item{direction}{A \code{\link{character}} string specifying the direction of the
relations in \code{x}. It must be one of "\code{above}" (the default) or "\code{below}"
(see details). Any unambiguous substring can be given.
Only used if \code{type} is "\code{stratigraphy}".}

\item{simplify}{A \code{\link{logical}} scalar: should multiple edges and loop edges
be removed?}

\item{reduce}{A \code{\link{logical}} scalar: should transitive reduction be performed?
Only used if \code{type} is "\code{stratigraphy}".}
}
\value{
An \pkg{igraph} graph object.
}
\description{
Creates an interval or a stratigraphic graph.
}
\details{
\describe{
\item{\code{interval}}{An interval graph is the graph showing intersecting
intervals on a line. As time is linear and not circular, an interval graph
contains no cycles with more than three edges and no shortcuts (it must be
a chordal graph).}
\item{\code{stratigraphy}}{A stratigraphic graph represents directed
relationships between (stratigraphic) units (it must be a directed acyclic
graph).}
}
}
\note{
Experimental: might change in a future release.

The \pkg{igraph} and \pkg{relations} packages needs to be installed on your
machine.
}
\examples{
if (requireNamespace("igraph", quietly = TRUE) &&
    requireNamespace("relations", quietly = TRUE)) {
  ## Seven intervals
  int <- intervals(
    start = c(1, 2, 3, 6, 9, 13, 17),
    end = c(7, 4, 15, 14, 11, 18, 19),
    calendar = CE(),
    names = c("A", "B", "C", "D", "E", "F", "G")
  )

  ## Interval graph
  g <- graph(int, type = "interval")
  plot(g)

  ## Stratigraphic graph
  g <- graph(int, type = "strati")
  plot(g, layout = igraph::layout_with_sugiyama)
}
}
\seealso{
Other chronological reasoning tools: 
\code{\link{overlap}()},
\code{\link{relations}}
}
\author{
N. Frerebeau
}
\concept{chronological reasoning tools}
