% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/calendar.R
\docType{methods}
\name{gregorian}
\alias{gregorian}
\alias{BP}
\alias{b2k}
\alias{BC}
\alias{BCE}
\alias{AD}
\alias{CE}
\title{Gregorian Calendar}
\usage{
BP(...)

b2k(...)

BC(...)

BCE(...)

AD(...)

CE(...)
}
\arguments{
\item{...}{Currently not used.}
}
\value{
A \code{\linkS4class{GregorianCalendar}} object.
}
\description{
Gregorian Calendar
}
\section{Functions}{
\itemize{
\item \code{BP()}: Gregorian BP era.

\item \code{b2k()}: Gregorian b2k era.

\item \code{BC()}: Gregorian BC era.

\item \code{BCE()}: Gregorian BCE era.

\item \code{AD()}: Gregorian AD era.

\item \code{CE()}: Gregorian CE era.

}}
\examples{
## Define time scales
calendar("BP")
calendar("AD")
calendar("julian")

## Shortcuts
BP()
AD()
J()
}
\seealso{
\code{\link[=calendar]{calendar()}}

Other calendar tools: 
\code{\link{calendar}()},
\code{\link{calendar_get}},
\code{\link{convert}()},
\code{\link{is}()},
\code{\link{julian}()}
}
\author{
N. Frerebeau
}
\concept{calendar tools}
