% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{GregorianCalendar-class}
\alias{GregorianCalendar-class}
\alias{.GregorianCalendar}
\title{GregorianCalendar}
\description{
An S4 class to represent a Gregorian calendar.
}
\seealso{
Other classes: 
\code{\link{AD-class}},
\code{\link{BC-class}},
\code{\link{BCE-class}},
\code{\link{BP-class}},
\code{\link{CE-class}},
\code{\link{JulianCalendar-class}},
\code{\link{RataDie-class}},
\code{\link{TimeLine-class}},
\code{\link{TimeScale-class}},
\code{\link{TimeSeries-class}},
\code{\link{b2k-class}}

Other calendar classes: 
\code{\link{JulianCalendar-class}},
\code{\link{TimeScale-class}}

Other gregorian era classes: 
\code{\link{AD-class}},
\code{\link{BC-class}},
\code{\link{BCE-class}},
\code{\link{BP-class}},
\code{\link{CE-class}},
\code{\link{b2k-class}}
}
\author{
N. Frerebeau
}
\concept{calendar classes}
\concept{classes}
\concept{gregorian era classes}
\keyword{internal}
