% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/interval.R
\docType{methods}
\name{hpdi}
\alias{hpdi}
\alias{hpdi-method}
\alias{hpdi,numeric,numeric-method}
\title{HPD Regions}
\usage{
hpdi(object, density, ...)

\S4method{hpdi}{numeric,numeric}(object, density, level = 0.954)
}
\arguments{
\item{object}{A \code{\link{numeric}} vector giving the coordinates of the points where
the density is estimated.}

\item{density}{A \code{\link{numeric}} vector giving the estimated density values.
If \code{density} is missing and \code{object} is a \code{ǹumeric} vector, density
estimates will be computed from \code{object}.}

\item{...}{Currently not used.}

\item{level}{A length-one \code{\link{numeric}} vector giving the confidence level.}
}
\value{
A \code{\link{list}} of \code{numeric} \code{\link{matrix}} giving the lower and upper boundaries of
the HPD interval and associated probabilities.
}
\description{
HPD Regions
}
\references{
Hyndman, R. J. (1996). Computing and graphing highest density regions.
\emph{American Statistician}, 50: 120-126. \doi{10.2307/2684423}.
}
\author{
N. Frerebeau
}
\concept{statistics}
