% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/interval.R
\docType{methods}
\name{interval_hdr}
\alias{interval_hdr}
\alias{interval_hdr-method}
\alias{interval_hdr,numeric,numeric-method}
\alias{interval_hdr,numeric,missing-method}
\title{HPD Regions}
\usage{
interval_hdr(x, y, ...)

\S4method{interval_hdr}{numeric,numeric}(x, y, level = 0.954)

\S4method{interval_hdr}{numeric,missing}(x, level = 0.954, ...)
}
\arguments{
\item{x}{A \code{\link{numeric}} vector giving the coordinates of the points where
the density is estimated.}

\item{y}{A \code{\link{numeric}} vector giving the estimated density values.
If \code{y} is missing and \code{x} is a \code{ǹumeric} vector, density estimates will be
computed from \code{x}.}

\item{...}{Currently not used.}

\item{level}{A length-one \code{\link{numeric}} vector giving the confidence level.}
}
\value{
A three-columns \code{numeric} \code{\link{matrix}} giving the lower and upper boundaries
of the HPD interval and associated probabilities.
}
\description{
HPD Regions
}
\references{
Hyndman, R. J. (1996). Computing and graphing highest density regions.
\emph{American Statistician}, 50: 120-126. \doi{10.2307/2684423}.
}
\author{
N. Frerebeau
}
\concept{statistics}
