% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/calendar-gregorian.R,
%   R/calendar-julian.R
\docType{methods}
\name{is}
\alias{is}
\alias{is_gregorian}
\alias{is_gregorian-method}
\alias{is_julian}
\alias{is_julian-method}
\alias{is_gregorian,ANY-method}
\alias{is_julian,ANY-method}
\title{Is an Object a Calendar?}
\usage{
is_gregorian(object)

is_julian(object)

\S4method{is_gregorian}{ANY}(object)

\S4method{is_julian}{ANY}(object)
}
\arguments{
\item{object}{Any \R object.}
}
\value{
A \code{\link{logical}} vector.
}
\description{
Test inheritance relationships between an object and a calendar class.
}
\seealso{
Other calendar tools: 
\code{\link{calendar_get}},
\code{\link{calendar}()},
\code{\link{convert}()},
\code{\link{gregorian}},
\code{\link{julian}()}
}
\author{
N. Frerebeau
}
\concept{calendar tools}
