% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\docType{methods}
\name{is_calendar}
\alias{is_calendar}
\alias{is_gregorian}
\alias{is_julian}
\title{Is an Object a Calendar?}
\usage{
is_calendar(object)

is_gregorian(object)

is_julian(object)
}
\arguments{
\item{object}{Any \R object.}
}
\value{
A \code{\link{logical}} scalar.
}
\description{
Test inheritance relationships between an object and a calendar class.
}
\seealso{
Other calendar tools: 
\code{\link{calendar}()},
\code{\link{calendar_get}},
\code{\link{convert}()},
\code{\link{get_calendar}()},
\code{\link{gregorian}},
\code{\link{julian}()}
}
\author{
N. Frerebeau
}
\concept{calendar tools}
