% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/calendar-julian.R
\docType{methods}
\name{julian}
\alias{julian}
\alias{is_julian}
\alias{is_julian-method}
\alias{is_julian,TimeScale-method}
\title{Julian Calendar}
\usage{
is_julian(object)

\S4method{is_julian}{TimeScale}(object)
}
\arguments{
\item{object}{A \code{\linkS4class{JulianCalendar}} object.}

\item{label}{A \code{\link{character}} string specifying the abbreviated label of
the time scale.}

\item{name}{A \code{\link{character}} string specifying the name of the time scale.}

\item{epoch}{A length-one \code{\link{numeric}} vector specifying the epoch year from
which years are counted (in Gregorian astronomical years).}

\item{direction}{A length-one \code{\link{integer}} vector specifying if years are
counted backwards (\code{-1}) or forwards (\code{1}) from \code{epoch}. Only the
\link[=sign]{sign} of \code{direction} will be retained.}
}
\value{
\itemize{
\item \code{as_julian()} returns a \code{\linkS4class{JulianCalendar}} object.
\item \code{is_julian()} returns a \code{\link{logical}} scalar.
}
}
\description{
Julian Calendar
}
\examples{
## Create a custom gregorian calendar
gregorian(
  label = "AUC",
  name = "Ab urbe condita",
  epoch = 753,
  direction = 1
)
}
\seealso{
Other time scales tools: 
\code{\link{convert}()},
\code{\link{era_get}},
\code{\link{era}()},
\code{\link{gregorian}}
}
\author{
N. Frerebeau
}
\concept{time scales tools}
