% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{JulianCalendar-class}
\alias{JulianCalendar-class}
\alias{.JulianCalendar}
\title{JulianCalendar}
\description{
An S4 class to represent a Julian calendar.
}
\seealso{
Other classes: 
\code{\link{Calendar-class}},
\code{\link{GregorianCalendar-class}},
\code{\link{TimeLine-class}},
\code{\link{TimeSeries-class}}

Other calendar classes: 
\code{\link{Calendar-class}},
\code{\link{GregorianCalendar-class}}
}
\author{
N. Frerebeau
}
\concept{calendar classes}
\concept{classes}
\keyword{internal}
