% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/mutators.R
\docType{methods}
\name{labels}
\alias{labels}
\alias{labels,TimeSeries-method}
\alias{labels,TimeIntervals-method}
\title{Labels}
\usage{
\S4method{labels}{TimeSeries}(object, ...)

\S4method{labels}{TimeIntervals}(object, ...)
}
\arguments{
\item{object}{An object.}

\item{...}{Currently not used.}
}
\value{
A \code{\link{character}} vector.
}
\description{
Find a suitable set of labels from an object.
}
\seealso{
Other mutators: 
\code{\link{as.data.frame}()},
\code{\link{subset}()}
}
\author{
N. Frerebeau
}
\concept{mutators}
