% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/mutators.R
\docType{methods}
\name{length}
\alias{length}
\alias{length,TimeIntervals-method}
\title{Length}
\usage{
\S4method{length}{TimeIntervals}(x)
}
\arguments{
\item{x}{An \R object.}
}
\value{
A length-one \code{\link{integer}} vector.
}
\description{
Get the length of an object.
}
\seealso{
Other mutators: 
\code{\link{as.data.frame}()},
\code{\link{labels}()},
\code{\link{names}()},
\code{\link{subset}()}
}
\author{
N. Frerebeau
}
\concept{mutators}
