% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/statistics.R
\docType{methods}
\name{mean}
\alias{mean}
\alias{mean,CalibratedAges-method}
\alias{mean,CalibratedAges,missing-method}
\title{Mean}
\usage{
\S4method{mean}{CalibratedAges}(x, na.rm = FALSE, ...)
}
\arguments{
\item{x}{A \code{\linkS4class{CalibratedAges}} object.}

\item{na.rm}{A \code{\link{logical}} scalar: should \code{NA} values be stripped before the
computation proceeds?}

\item{...}{Currently not used.}
}
\description{
Mean
}
\examples{
## Calibrate a single date
cal <- c14_calibrate(130, 20)

## Statistics
median(cal)
mean(cal)

## Plot
plot(cal, panel.first = graphics::grid())
abline(v = median(cal), lty = 2, col = "blue")
abline(v = mean(cal), lty = 2, col = "red")
}
\seealso{
Other statistics: 
\code{\link{hpdi}()},
\code{\link{median}()}
}
\author{
N. Frerebeau
}
\concept{statistics}
