% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/mutators.R
\docType{methods}
\name{names}
\alias{names}
\alias{names,TimeIntervals-method}
\alias{names<-,TimeIntervals-method}
\title{Names}
\usage{
\S4method{names}{TimeIntervals}(x)

\S4method{names}{TimeIntervals}(x) <- value
}
\arguments{
\item{x}{An \R object.}

\item{value}{A \code{\link{character}} vector of up to the same length as \code{x}, or
\code{NULL}.}
}
\value{
The updated object.
}
\description{
Get or set the names of an object.
}
\seealso{
Other mutators: 
\code{\link{as.data.frame}()},
\code{\link{labels}()},
\code{\link{length}()},
\code{\link{subset}()}
}
\author{
N. Frerebeau
}
\concept{mutators}
