% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/show.R
\docType{methods}
\name{pretty}
\alias{pretty}
\alias{pretty,RataDie-method}
\title{Pretty Breakpoints}
\usage{
\S4method{pretty}{RataDie}(x, calendar = getOption("chronos.calendar"), ...)
}
\arguments{
\item{x}{A \code{\linkS4class{RataDie}} object.}

\item{calendar}{A \code{\linkS4class{TimeScale}} object specifying the target calendar
(see \code{\link[=calendar]{calendar()}}).}

\item{...}{Further parameters to be passed to \code{\link[base:pretty]{base::pretty()}}.}
}
\value{
A \code{\linkS4class{RataDie}} object.
}
\description{
Pretty Breakpoints
}
\details{
\code{pretty()} computes a vector of increasing numbers which are "pretty" in
decimal notation of \code{calendar}. Pretty breakpoints are then converted to
\emph{rata die}.
}
\seealso{
Other fixed date tools: 
\code{\link{arithmetic}},
\code{\link{as_date}()},
\code{\link{as_decimal}()},
\code{\link{as_year}()},
\code{\link{fixed_gregorian}},
\code{\link{fixed_julian}},
\code{\link{fixed}()},
\code{\link{format}()}
}
\concept{fixed date tools}
