% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/era.R
\docType{methods}
\name{project}
\alias{project}
\alias{project-method}
\alias{project,CalibratedAges,character-method}
\alias{project,CalibratedSPD,character-method}
\title{Change the Time Scale}
\usage{
project(object, target, ...)

\S4method{project}{CalibratedAges,character}(object, target)

\S4method{project}{CalibratedSPD,character}(object, target)
}
\arguments{
\item{object}{An object describing the source era.}

\item{target}{A \code{\link{character}} string specifying the target time scale
(see \code{\link[=era]{era()}}).}

\item{...}{Currently not used.}
}
\value{
An object.
}
\description{
Change the time scale of an object.
}
\seealso{
Other time scales: 
\code{\link{convert}()},
\code{\link{era}()},
\code{\link{years}()}
}
\author{
N. Frerebeau
}
\concept{time scales}
