% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{RataDie-class}
\alias{RataDie-class}
\alias{.RataDie}
\title{RataDie}
\description{
An S4 class to represent a vector of \emph{rata die}.
}
\details{
\emph{Rata die} (fixed date) are represented as the number of days since 01-01-01
(Gregorian), with negative values for earlier dates.

It is intended that the date should be an integer value, but this is not
enforced in the internal representation.

When printing there is assumed to be a year zero.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{A \code{\link{numeric}} vector giving the \emph{rata die} values.}
}}

\note{
This class inherits from \code{\link{numeric}}.
}
\seealso{
Other classes: 
\code{\link{AD-class}},
\code{\link{BC-class}},
\code{\link{BCE-class}},
\code{\link{BP-class}},
\code{\link{CE-class}},
\code{\link{GregorianCalendar-class}},
\code{\link{JulianCalendar-class}},
\code{\link{TimeScale-class}},
\code{\link{TimeSeries-class}},
\code{\link{b2k-class}}

Other time classes: 
\code{\link{TimeSeries-class}}
}
\author{
N. Frerebeau
}
\concept{classes}
\concept{time classes}
\keyword{internal}
