% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhythm.R
\name{rhythm}
\alias{rhythm}
\title{Rhythm of events}
\usage{
rhythm(x, t, ...)
}
\arguments{
\item{x}{Vector of potential data points that reflect (or do not reflect) event data}

\item{t}{Vector of time points that reflect information about events}
}
\value{
A "combined" record of the \code{rhythm} class that has validated event
and time properties.
}
\description{
The type of data that internally helps to parse data will be built and
validated on this class definition. This allos for understanding of the type
of variable (continuous or categorical) and how it is sampled (regular or
irregular intervals).
}
\details{
Both \code{x} and \code{t} must be vectors of the same length.
}
