% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/radiocarbon.R
\docType{methods}
\name{spd}
\alias{spd}
\alias{spd-method}
\alias{spd,CalibratedAges-method}
\title{Summed Probability Distributions}
\usage{
spd(object, ...)

\S4method{spd}{CalibratedAges}(object, normalize_date = FALSE, normalize_spd = FALSE)
}
\arguments{
\item{object}{A \code{\linkS4class{CalibratedAges}} object.}

\item{...}{Currently not used.}

\item{normalize_date}{A \code{\link{logical}} scalar: should the total probability mass
of the calibrated dates be normalised (to sum to unity within the time-span
of analysis)?}

\item{normalize_spd}{A \code{\link{logical}} scalar: should the total probability mass
of the SPD be normalised (to sum to unity)?}
}
\description{
Computes summed probability distributions (SPD) of radiocarbon dates.
}
\examples{
## Radiocarbon data from Bosch et al. 2015
data("ksarakil")

## Calibrate
cal <- c14_calibrate(
  ages = ksarakil$date,
  errors = ksarakil$error,
  names = ksarakil$code,
  curves = "marine13",
  reservoir_offsets = 53,
  reservoir_errors = 43,
  from = 50000, to = 0
)
plot(cal)

## SPD
s <- spd(cal)
plot(s)
}
\seealso{
Other radiocarbon tools: 
\code{\link{F14C}},
\code{\link{c14_calibrate}()},
\code{\link{c14_combine}()},
\code{\link{c14_curve}()},
\code{\link{c14_plot}}
}
\author{
N. Frerebeau
}
\concept{radiocarbon tools}
