% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/series.R
\docType{methods}
\name{start}
\alias{start}
\alias{start-method}
\alias{end-method}
\alias{start,TimeSeries-method}
\alias{end,TimeSeries-method}
\title{Terminal Times of Time Series}
\usage{
\S4method{start}{TimeSeries}(x)

\S4method{end}{TimeSeries}(x)
}
\arguments{
\item{x}{A \code{\linkS4class{TimeSeries}} object.}
}
\value{
A \code{\link{numeric}} vector.
}
\description{
Get the times the first and last observations were taken.
}
\examples{
## Create time-series of 100 observations
## Sampled every years starting from 1000 CE

## Univariate
X <- series(rnorm(300), era("BCE"), start = 1000)

names(X)
start(X)
end(X)
time(X)

## Multivariate
Y <- series(matrix(rnorm(300), 100, 3), era("CE"), start = 1000)

names(Y)
start(Y)
end(Y)
time(Y)
}
\seealso{
Other time series tools: 
\code{\link{plot}()},
\code{\link{project}()},
\code{\link{series}()},
\code{\link{time}()},
\code{\link{window}()}
}
\author{
N. Frerebeau
}
\concept{time series tools}
