% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/subset.R
\docType{methods}
\name{subset}
\alias{subset}
\alias{[,RataDie,ANY,ANY,ANY-method}
\alias{[,RataDie-method}
\alias{[,TimeSeries,ANY,ANY,ANY-method}
\alias{[,TimeSeries-method}
\title{Extract or Replace Parts of an Object}
\usage{
\S4method{[}{RataDie,ANY,ANY,ANY}(x, i)

\S4method{[}{TimeSeries,ANY,ANY,ANY}(x, i, j, k, drop = FALSE)
}
\arguments{
\item{x}{An object from which to extract element(s) or in which to replace
element(s).}

\item{i, j, k}{Indices specifying elements to extract or replace.}

\item{drop}{A \code{\link{logical}} scalar: should the result be coerced to
the lowest possible dimension? This only works for extracting elements,
not for the replacement.}
}
\value{
A subsetted object.
}
\description{
Operators acting on objects to extract or replace parts.
}
\author{
N. Frerebeau
}
\concept{mutators}
