% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{TimeLine-class}
\alias{TimeLine-class}
\alias{.TimeLine}
\title{TimeLine}
\description{
An S4 class to represent a vector of years.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{A \code{\link{numeric}} vector giving the year values.}

\item{\code{error}}{A \code{\link{numeric}} vector of uncertainties.}

\item{\code{scale}}{A \code{\link{numeric}} value specifying the number of years represented by
one unit. It should be a power of 10 (i.e. 1000 means ka).}

\item{\code{calendar}}{A \code{\linkS4class{Calendar}} object specifying the time scale.}
}}

\note{
This class inherits from \code{\link{numeric}}.
}
\seealso{
Other classes: 
\code{\link{Calendar-class}},
\code{\link{GregorianCalendar-class}},
\code{\link{JulianCalendar-class}},
\code{\link{TimeSeries-class}}

Other time classes: 
\code{\link{TimeSeries-class}}
}
\author{
N. Frerebeau
}
\concept{classes}
\concept{time classes}
\keyword{internal}
