% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{TimeScale-class}
\alias{TimeScale-class}
\alias{.TimeScale}
\title{TimeScale}
\description{
An S4 class to represent a time scale.
}
\section{Slots}{

\describe{
\item{\code{label}}{A \code{\link{character}} string giving the abbreviated label of the time
scale.}

\item{\code{name}}{A \code{\link{character}} string giving the name of the time scale.}

\item{\code{epoch}}{A \code{\link{numeric}} value giving the epoch year from which years are
counted (in Gregorian astronomical years).}

\item{\code{scale}}{An \code{\link{integer}} specifying the number of years represented by one
unit.}

\item{\code{direction}}{An \code{\link{integer}} specifying if years are counted backwards
(\eqn{-1}) or forwards (\eqn{1}) from \code{epoch}.}
}}

\note{
This class inherits from \code{\linkS4class{Calendar}}.
}
\seealso{
Other classes: 
\code{\link{Calendar-class}}
}
\author{
N. Frerebeau
}
\concept{classes}
\keyword{internal}
