% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{TimeSeries-class}
\alias{TimeSeries-class}
\alias{.TimeSeries}
\title{TimeSeries}
\description{
An S4 class to represent time series.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{A \code{numeric} \code{\link{matrix}} giving the observed time-series values.}

\item{\code{time}}{A \code{\linkS4class{RataDie}} object.}
}}

\note{
This class inherits from \code{\link{matrix}}.
}
\seealso{
Other classes: 
\code{\link{AD-class}},
\code{\link{BC-class}},
\code{\link{BCE-class}},
\code{\link{BP-class}},
\code{\link{CE-class}},
\code{\link{GregorianCalendar-class}},
\code{\link{JulianCalendar-class}},
\code{\link{RataDie-class}},
\code{\link{TimeScale-class}},
\code{\link{b2k-class}}

Other time classes: 
\code{\link{RataDie-class}}
}
\author{
N. Frerebeau
}
\concept{classes}
\concept{time classes}
\keyword{internal}
