% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plot.R
\docType{methods}
\name{year_grid}
\alias{year_grid}
\title{Add Grid to a Plot}
\usage{
year_grid(
  nx = NULL,
  ny = NULL,
  col = "lightgray",
  lty = "dotted",
  lwd = graphics::par("lwd"),
  calendar = getOption("aion.last_calendar")
)
}
\arguments{
\item{nx, ny}{A \code{\link{numeric}} value giving the number of cells of the grid in
\code{x} and \code{y} direction. When \code{NULL}, as per default, the grid aligns with the
tick marks on the corresponding default axis (i.e., tickmarks as computed by
\code{\link[=year_axis]{year_axis()}}). When \code{NA}, no grid lines are drawn in the corresponding
direction.}

\item{col}{A \code{\link{character}} string specifying the color of the grid lines.}

\item{lty}{A \code{\link{character}} string or numeric value specifying the line type
of the grid lines.}

\item{lwd}{A non-negative \code{\link{numeric}} value specifying the line width of the
grid lines.}

\item{calendar}{A \code{\linkS4class{TimeScale}} object specifying the target calendar
(see \code{\link[=calendar]{calendar()}}).}
}
\value{
\code{year_grid()} is called it for its side-effects.
}
\description{
Add Grid to a Plot
}
\examples{
## Create a time-series of 300 observations
## Sampled every two years starting from 2000 BP
X <- series(
  object = rnorm(300),
  time = seq(2000, by = 2, length.out = 300),
  calendar = BP()
)

## Axis
plot(X, axes = FALSE, calendar = BP()) # Remove axes
year_axis(X, side = 1) # Same calendar as last plot
year_axis(X, side = 3, calendar = CE()) # Specific calendar
mtext(format(CE()), side = 3, line = 3)

## Grid
## grid() does not align with the tick marks of the default axis
## because x coordinates of the plotting region are expressed in rata die
## but year_axis() computes tickmarks according to a given calendar
plot(X, panel.first = graphics::grid())

plot(X, panel.first = year_grid(12))
plot(X, panel.first = year_grid(12, NA))
}
\seealso{
Other plotting tools: 
\code{\link{image}()},
\code{\link{plot}()},
\code{\link{year_axis}()}
}
\author{
N. Frerebeau
}
\concept{plotting tools}
