% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/era.R
\docType{methods}
\name{years}
\alias{years}
\alias{years-method}
\alias{years,CalibratedAges-method}
\alias{years,CalibratedSPD-method}
\title{Time Grid}
\usage{
years(object)

\S4method{years}{CalibratedAges}(object)

\S4method{years}{CalibratedSPD}(object)
}
\arguments{
\item{object}{An object.}
}
\value{
A \code{\link{numeric}} vector.
}
\description{
Get the time grid of an object.
}
\seealso{
Other time scales: 
\code{\link{convert}()},
\code{\link{era}()},
\code{\link{project}()}
}
\author{
N. Frerebeau
}
\concept{time scales}
