% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aion-package.R
\docType{package}
\name{aion-package}
\alias{aion-package}
\alias{aion}
\title{aion: Archaeological Time Series}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A toolkit for archaeological time series and time intervals. This package provides a system of classes and methods to represent and work with archaeological time series and time intervals. Dates are represented as "rata die" and can be converted to (virtually) any calendar defined by Reingold and Dershowitz (2018) \doi{10.1017/9781107415058}. This packages offers a simple API that can be used by other specialized packages.
}
\details{
\tabular{ll}{
\strong{Version} \tab 1.6.0 \cr
\strong{License} \tab GPL-3 \cr
\strong{CRAN DOI} \tab \doi{10.32614/cran.package.aion} \cr
\strong{Zenodo DOI} \tab \doi{10.5281/zenodo.8032278} \cr
\strong{JOSS DOI} \tab \doi{10.21105/joss.06210} \cr
}

Archéosciences Bordeaux (UMR 6034)\cr
Maison de l'Archéologie\cr
Université Bordeaux Montaigne\cr
F-33607 Pessac cedex\cr
France
}
\section{Package options}{

\pkg{aion} uses the following \code{\link[=options]{options()}} to configure behaviour:
\itemize{
\item \code{aion.verbose}: a \code{\link{logical}} scalar. Should \R report extra information
on progress? Defaults to \code{\link[=interactive]{interactive()}}.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://codeberg.org/tesselle/aion}
  \item \url{https://tesselle.r-universe.dev/aion}
  \item \url{https://packages.tesselle.org/aion/}
  \item Report bugs at \url{https://codeberg.org/tesselle/aion/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Frerebeau \email{nicolas.frerebeau@u-bordeaux-montaigne.fr} (\href{https://orcid.org/0000-0001-5759-4944}{ORCID})

Authors:
\itemize{
  \item Joe Roe \email{joseph.roe@iaw.unibe.ch} (\href{https://orcid.org/0000-0002-1011-1244}{ORCID})
}

Other contributors:
\itemize{
  \item Brice Lebrun (\href{https://orcid.org/0000-0001-7503-8685}{ORCID}) (Logo designer) [artist]
  \item Université Bordeaux Montaigne (\href{https://ror.org/03pbgwk21}{ROR}) [funder]
  \item CNRS (\href{https://ror.org/02feahw73}{ROR}) [funder]
}

}
\keyword{internal}
