% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/allen.R
\docType{methods}
\name{allen_complement}
\alias{allen_complement}
\alias{allen_complement-method}
\alias{allen_complement,character-method}
\alias{allen_complement,matrix-method}
\title{Complement of an Allen Relation}
\usage{
allen_complement(x, ...)

\S4method{allen_complement}{character}(x)

\S4method{allen_complement}{matrix}(x)
}
\arguments{
\item{x}{A \code{\link{character}} vector or matrix of Allen relations (typically
returned by \code{\link[=allen_relation]{allen_relation()}}).}

\item{...}{Currently not used.}
}
\value{
A \code{\link{character}} vector or matrix (same as \code{x}).
}
\description{
Complement of an Allen Relation
}
\examples{
## Data from Husi 2022
data("loire", package = "folio")
loire <- subset(loire, area == "Anjou")

## Basic relations
allen_relation(loire$lower, loire$upper)

## Complement
(comp <- allen_complement("F")) # "pmoDseSdfOMP"

## Converse
(conv <- allen_converse(comp)) # "pmoFDseSdOMP"

## Composition
allen_composition("oFD", "oFDseS") # "pmoFD"

## Intersection
allen_intersect("pFsSf", "pmoFD") # "pF"

# Union
allen_union("pFsSf", "pmoFD") # "pmoFDsSf"
}
\references{
Allen, J. F. (1983). Maintaining Knowledge about Temporal Intervals.
\emph{Communications of the ACM}, 26(11): 832-843. \doi{10.1145/182.358434}.
}
\seealso{
Other Allen's intervals: 
\code{\link{allen_composition}()},
\code{\link{allen_converse}()},
\code{\link{allen_intersect}()},
\code{\link{allen_relation}()},
\code{\link{allen_union}()}
}
\author{
N. Frerebeau
}
\concept{Allen's intervals}
