% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/allen.R
\docType{methods}
\name{allen_union}
\alias{allen_union}
\alias{allen_union-method}
\alias{allen_union,character,character-method}
\title{Union of Allen Relations}
\usage{
allen_union(x, y, ...)

\S4method{allen_union}{character,character}(x, y)
}
\arguments{
\item{x, y}{A \code{\link{character}} vector of Allen relations.}

\item{...}{Currently not used.}
}
\value{
A \code{\link{character}} vector.
}
\description{
Union of Allen Relations
}
\examples{
## Seven intervals
int <- intervals(
  start = c(1, 2, 3, 6, 9, 13, 17),
  end = c(7, 4, 15, 14, 11, 18, 19),
  calendar = CE(),
  names = c("A", "B", "C", "D", "E", "F", "G")
)

## Basic relations
allen_relation(int)

## Complement
(comp <- allen_complement("F"))

## Converse
(conv <- allen_converse(comp))

## Composition
allen_composition("oFD", "oFDseS")

## Intersection
allen_intersect("pFsSf", "pmoFD")

# Union
allen_union("pFsSf", "pmoFD")
}
\references{
Allen, J. F. (1983). Maintaining Knowledge about Temporal Intervals.
\emph{Communications of the ACM}, 26(11): 832-843. \doi{10.1145/182.358434}.
}
\seealso{
Other Allen's intervals: 
\code{\link{allen_complement}()},
\code{\link{allen_composition}()},
\code{\link{allen_converse}()},
\code{\link{allen_intersect}()},
\code{\link{allen_relation}()},
\code{\link{allen_relation_code}()}
}
\author{
T. S. Dye, N. Frerebeau
}
\concept{Allen's intervals}
