% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/operators.R
\docType{methods}
\name{arithmetic}
\alias{arithmetic}
\alias{Arith,RataDie,RataDie-method}
\title{Arithmetic Operators}
\usage{
\S4method{Arith}{RataDie,RataDie}(e1, e2)
}
\arguments{
\item{e1, e2}{A \code{\linkS4class{RataDie}} object.}
}
\value{
A \code{\link{logical}} vector.
}
\description{
Operators performing chronological comparison (not numerical).
}
\details{
Years will be coerced to a plain numeric vector if a computation means their
era no longer makes sense.
}
\examples{
## Vectors of years
x <- as_fixed(c(-350, 31, 1072, 576, 1130), calendar = CE())
y <- as_fixed(c(1494, 1645, -869, 1440, 1851), calendar = CE())

## Move forward in time
x + y

## Move backward in time
x - y
}
\seealso{
Other operators: 
\code{\link{compare}}
}
\author{
N. Frerebeau
}
\concept{operators}
