% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/operators.R
\docType{methods}
\name{arithmetic}
\alias{arithmetic}
\alias{Arith,TimeLine,TimeLine-method}
\title{Arithmetic Operators}
\usage{
\S4method{Arith}{TimeLine,TimeLine}(e1, e2)
}
\arguments{
\item{e1, e2}{A \code{\linkS4class{TimeLine}} object.}
}
\value{
A \code{\link{logical}} vector.
}
\description{
Operators performing chronological comparison (not numerical).
}
\details{
The following operators allow to perform artithmetic operations on years,
\strong{relatively to the epoch of the era} (this ensures that the results are
the same, regardless of the era).

\tabular{ll}{
\strong{operator} \tab \strong{meaning} \cr
\code{+}          \tab move forward in time \cr
\code{-}          \tab move backward in time \cr
}

Years will be coerced to a plain numeric vector if a computation means their
era no longer makes sense.
}
\examples{
## /!\ Be carefull: computations are relative to the epoch /!\

## Vectors of years
(x_CE <- years(c(-350, 31, 1072, 576, 1130), calendar("CE")))
(y_CE <- years(c(1494, 1645, -869, 1440, 1851), calendar("CE")))

## Move forward in time
x_CE + y_CE

## Move backward in time
x_CE - y_CE

## Change the origin
(x_BP <- project(x_CE, calendar("BP")))
(y_BP <- project(y_CE, calendar("BP")))

## Move forward in time
(z_BP <- x_BP + y_BP)

## Change the origin back to CE
(z_CE <- project(z_BP, calendar("CE")))

all.equal(x_CE + y_CE, z_CE)
}
\seealso{
Other operators: 
\code{\link{compare}}
}
\author{
N. Frerebeau
}
\concept{operators}
