% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/calendar-gregorian.R
\docType{methods}
\name{as_date}
\alias{as_date}
\alias{as_date-method}
\alias{as_date,numeric,GregorianCalendar-method}
\title{Date Conversion from \emph{Rata Die}}
\usage{
as_date(object, calendar)

\S4method{as_date}{numeric,GregorianCalendar}(object, calendar)
}
\arguments{
\item{object}{A \code{\linkS4class{RataDie}} object (see \code{\link[=fixed]{fixed()}}).}

\item{calendar}{A \code{\linkS4class{TimeScale}} object specifying the target calendar
(see \code{\link[=calendar]{calendar()}}).}
}
\value{
A \code{\link{numeric}} vector of (decimal) years.
}
\description{
Date Conversion from \emph{Rata Die}
}
\examples{
## R 1.0.0
(y <- fixed(year = 2000, month = 02, day = 29, calendar = CE()))
as_date(y, calendar = CE())
as_year(y, calendar = CE())

## Vector of years expressed in ka BP
(ka <- fixed(c(30, 35, 40), calendar = BP(), scale = 1000))
as_year(ka, calendar = CE())
}
\references{
Reingold, E. M. and Dershowitz, N. (2018). \emph{Calendrical Calculations:
The Ultimate Edition}. Cambridge University Press.
\doi{10.1017/9781107415058}.
}
\seealso{
Other fixed date tools: 
\code{\link{as_decimal}()},
\code{\link{as_year}()},
\code{\link{fixed_gregorian}},
\code{\link{fixed}()}
}
\author{
N. Frerebeau
}
\concept{fixed date tools}
