% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/calendar-gregorian.R
\docType{methods}
\name{as_date}
\alias{as_date}
\alias{as_date-method}
\alias{as_date,numeric,GregorianCalendar-method}
\title{Date Conversion from Rata Die}
\usage{
as_date(object, calendar)

\S4method{as_date}{numeric,GregorianCalendar}(object, calendar = calendar("CE"))
}
\arguments{
\item{object}{A \code{\linkS4class{RataDie}} object (see \code{\link[=as_fixed]{as_fixed()}}).}

\item{calendar}{A \code{\linkS4class{TimeScale}} object (see \code{\link[=calendar]{calendar()}}).}
}
\description{
Date Conversion from Rata Die
}
\examples{
## R foundation
(y <- years(c(1994, 2000), calendar("CE")))
format(y)

## Vector of years expressed in ka BP
(ka <- years(c(30, 35, 40), calendar("BP"), scale = 1000))
format(ka)
format(ka, format = "ka")
}
\seealso{
Other time series tools: 
\code{\link{as_fixed}()},
\code{\link{as_year}()},
\code{\link{plot}()},
\code{\link{project}()},
\code{\link{series}()},
\code{\link{start}()},
\code{\link{time}()},
\code{\link{window}()},
\code{\link{years}()}
}
\author{
N. Frerebeau
}
\concept{time series tools}
