% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/calendar-gregorian.R
\docType{methods}
\name{as_fixed}
\alias{as_fixed}
\alias{as_fixed-method}
\alias{as_fixed,numeric,missing,missing,GregorianCalendar-method}
\alias{as_fixed,numeric,numeric,numeric,GregorianCalendar-method}
\title{Rata Die}
\usage{
as_fixed(year, month, day, calendar)

\S4method{as_fixed}{numeric,missing,missing,GregorianCalendar}(year, calendar = calendar("CE"))

\S4method{as_fixed}{numeric,numeric,numeric,GregorianCalendar}(year, month, day, calendar = calendar("CE"))
}
\arguments{
\item{year}{A \code{\link{numeric}} vector of (decimal) years.}

\item{month}{A \code{\link{numeric}} vector of (decimal) years.}

\item{day}{A \code{\link{numeric}} vector of (decimal) years.}

\item{calendar}{A \code{\linkS4class{TimeScale}} object (see \code{\link[=calendar]{calendar()}}).}
}
\value{
A \code{\linkS4class{RataDie}} object.
}
\description{
Rata Die
}
\examples{
## R foundation
(y <- years(c(1994, 2000), calendar("CE")))
format(y)

## Vector of years expressed in ka BP
(ka <- years(c(30, 35, 40), calendar("BP"), scale = 1000))
format(ka)
format(ka, format = "ka")
}
\seealso{
Other time series tools: 
\code{\link{as_date}()},
\code{\link{as_year}()},
\code{\link{plot}()},
\code{\link{project}()},
\code{\link{series}()},
\code{\link{start}()},
\code{\link{time}()},
\code{\link{window}()},
\code{\link{years}()}
}
\author{
N. Frerebeau
}
\concept{time series tools}
