% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/calendar.R
\docType{methods}
\name{calendar}
\alias{calendar}
\alias{calendar-method}
\alias{calendar,character-method}
\alias{calendar,TimeLine-method}
\alias{calendar,TimeSeries-method}
\title{Era}
\usage{
calendar(object, ...)

\S4method{calendar}{character}(object)

\S4method{calendar}{TimeLine}(object)

\S4method{calendar}{TimeSeries}(object)
}
\arguments{
\item{object}{A \code{\link{character}} string specifying the abbreviated label of
the time scale (see details) or an object from which to extract the time
scale.}

\item{...}{Currently not used.}
}
\value{
A \code{\linkS4class{TimeScale}} object.
}
\description{
Era
}
\details{
The following time scales are available:

\tabular{lll}{
\strong{label}   \tab \strong{name}           \tab \strong{unit}        \cr
\code{BP}        \tab Before Present     \tab Gregorian years \cr
\code{BC}        \tab Before Christ      \tab Gregorian years \cr
\code{BCE}       \tab Before Common Era  \tab Gregorian years \cr
\code{AD}        \tab Anno Domini        \tab Gregorian years \cr
\code{CE}        \tab Common Era         \tab Gregorian years \cr
\code{b2k}       \tab Years before 2000  \tab Gregorian years \cr
}
}
\note{
Inspired by \code{\link[era:era]{era::era()}} by Joe Roe.
}
\examples{
## Define time scales
(BP <- calendar("BP"))
(AD <- calendar("AD"))

## Create a custom gregorian calendar
as_gregorian(
  label = "AUC",
  name = "Ab urbe condita",
  epoch = 753,
  direction = 1
)
}
\seealso{
Other time scales tools: 
\code{\link{calendar_get}},
\code{\link{convert}()},
\code{\link{gregorian}},
\code{\link{julian}()}
}
\author{
N. Frerebeau
}
\concept{time scales tools}
