% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/calendar.R
\docType{methods}
\name{calendar_get}
\alias{calendar_get}
\alias{calendar_label}
\alias{calendar_label-method}
\alias{calendar_name}
\alias{calendar_name-method}
\alias{calendar_unit}
\alias{calendar_unit-method}
\alias{calendar_epoch}
\alias{calendar_epoch-method}
\alias{calendar_direction}
\alias{calendar_direction-method}
\alias{calendar_label,TimeScale-method}
\alias{calendar_name,TimeScale-method}
\alias{calendar_unit,TimeScale-method}
\alias{calendar_epoch,TimeScale-method}
\alias{calendar_direction,TimeScale-method}
\title{Calendar Parameters}
\usage{
calendar_label(object)

calendar_name(object)

calendar_unit(object)

calendar_epoch(object)

calendar_direction(object)

\S4method{calendar_label}{TimeScale}(object)

\S4method{calendar_name}{TimeScale}(object)

\S4method{calendar_unit}{TimeScale}(object)

\S4method{calendar_epoch}{TimeScale}(object)

\S4method{calendar_direction}{TimeScale}(object)
}
\arguments{
\item{object}{A \code{\linkS4class{TimeScale}} object.}
}
\description{
Calendar Parameters
}
\examples{
## Define time scales
(BP <- calendar("BP"))
(AD <- calendar("AD"))
}
\seealso{
Other time scales tools: 
\code{\link{calendar}()},
\code{\link{convert}()},
\code{\link{gregorian}},
\code{\link{julian}()}
}
\author{
N. Frerebeau
}
\concept{time scales tools}
