% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CalibratedSPD-class}
\alias{CalibratedSPD-class}
\alias{.CalibratedSPD}
\title{Calibrated SPD}
\description{
An S4 class to represent summed probability distributions (SPD) of
radiocarbon dates.
}
\note{
This class inherits from \code{\linkS4class{TimeSeries}}.
}
\seealso{
Other classes: 
\code{\link{CalibratedAges-class}},
\code{\link{TimeScale-class}},
\code{\link{TimeSeries-class}}
}
\author{
N. Frerebeau
}
\concept{classes}
\keyword{internal}
