% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronos-package.R
\docType{package}
\name{chronos-package}
\alias{chronos-package}
\alias{_PACKAGE}
\alias{chronos}
\title{chronos: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\details{
\tabular{ll}{
\strong{Package:} \tab chronos \cr
\strong{Type:} \tab Package \cr
\strong{Version:} \tab 0.0.0 \cr
\strong{License:} \tab GPL-3 \cr
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://packages.tesselle.org/chronos/}
  \item \url{https://github.com/tesselle/chronos}
  \item Report bugs at \url{https://github.com/tesselle/chronos/issues}
}

}
\author{
\strong{Full list of authors and contributors} (alphabetic order):

\tabular{ll}{
Nicolas Frerebeau \tab \emph{Université Bordeaux Montaigne, France} \cr
Brice Lebrun \tab \emph{Université Bordeaux Montaigne, France} \cr
}

\strong{Package maintainer}

Nicolas Frerebeau\cr
\email{nicolas.frerebeau@u-bordeaux-montaigne.fr}

Archéosciences Bordeaux (UMR 6034)\cr
Maison de l'Archéologie\cr
Université Bordeaux Montaigne\cr
F-33607 Pessac cedex\cr
France
}
\keyword{internal}
