% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronos-package.R
\docType{package}
\name{chronos-package}
\alias{chronos-package}
\alias{_PACKAGE}
\alias{chronos}
\title{chronos: Archaeological Time Series}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A toolkit for archaeological time series and temporal reasoning. This package provides classes and methods to represent and work with archaeological time series. This package includes common time scales (e.g., before present, common era) and supports user-defined scales. Time series can be reprojected to a different time scale (e.g. from BP to CE years). This package includes common time scales (e.g., before present, common era) and supports user-defined scales. It also includes tools to describe and analyze finite time intervals.
}
\details{
\tabular{ll}{
\strong{Package:} \tab chronos \cr
\strong{Type:} \tab Package \cr
\strong{Version:} \tab 0.0.0 \cr
\strong{License:} \tab GPL-3 \cr
}
}
\section{Package options}{

\code{chronos} uses the following \code{\link[=options]{options()}} to configure behaviour:
\itemize{
\item \code{chronos.precision}: an \code{\link{integer}} indicating the number of decimal
places (defaults to \eqn{0}).
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://packages.tesselle.org/chronos/}
  \item \url{https://github.com/tesselle/chronos}
  \item Report bugs at \url{https://github.com/tesselle/chronos/issues}
}

}
\author{
\strong{Full list of authors and contributors} (alphabetic order):

\tabular{ll}{
Nicolas Frerebeau \tab \emph{Université Bordeaux Montaigne, France} \cr
Brice Lebrun \tab \emph{Université Bordeaux Montaigne, France} \cr
}

\strong{Package maintainer}

Nicolas Frerebeau\cr
\email{nicolas.frerebeau@u-bordeaux-montaigne.fr}

Archéosciences Bordeaux (UMR 6034)\cr
Maison de l'Archéologie\cr
Université Bordeaux Montaigne\cr
F-33607 Pessac cedex\cr
France
}
\keyword{internal}
