% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/operators.R
\docType{methods}
\name{compare}
\alias{compare}
\alias{Compare,RataDie,RataDie-method}
\title{Chronological Comparison}
\usage{
\S4method{Compare}{RataDie,RataDie}(e1, e2)
}
\arguments{
\item{e1, e2}{A \code{\linkS4class{RataDie}} object.}
}
\value{
A \code{\link{logical}} vector.
}
\description{
Operators performing chronological comparison (not numerical).
}
\details{
The following operators allow to compare two objects. Remember that this is
a \strong{chronological comparison} and not a numerical comparison (e.g. 5000 is
numerically greater than 3000, but 5000 BP precedes 3000 BP).

\tabular{ll}{
\strong{operator} \tab \strong{meaning} \cr
\code{==}         \tab equals \cr
\code{!=}         \tab differs \cr
\code{>}          \tab preceded by \cr
\code{<}          \tab precedes \cr
\code{>=}         \tab preceded by or equals \cr
\code{<=}         \tab precedes or equals \cr
}
}
\examples{
## Vectors of years
x <- fixed(c(-9551, -7917, -8387, -3371, -3004), calendar = CE())
y <- fixed(c(-3940, -4055, -8528, -5068, -8442), calendar = CE())

## x precedes y?
x < y

## x preceded by y?
x > y
}
\seealso{
Other operators: 
\code{\link{arithmetic}}
}
\author{
N. Frerebeau
}
\concept{operators}
