% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/project.R
\docType{methods}
\name{convert}
\alias{convert}
\alias{convert-method}
\alias{convert,character,character-method}
\alias{convert,Calendar,Calendar-method}
\title{Calendar Converter}
\usage{
convert(from, to, ...)

\S4method{convert}{character,character}(from, to)

\S4method{convert}{Calendar,Calendar}(from, to)
}
\arguments{
\item{from}{A \code{\linkS4class{Calendar}} object describing the source calendar.}

\item{to}{A \code{\linkS4class{Calendar}} object describing the target calendar.}

\item{...}{Currently not used.}
}
\value{
A \code{\link{function}} that when called with a single numeric argument (years)
converts years from one calendar to another.
}
\description{
Interconverts dates in a variety of calendars.
}
\note{
Adapted from \code{\link[era:yr_transform]{era::yr_transform()}} by Joe Roe.
}
\seealso{
Other time scales tools: 
\code{\link{as_calendar}},
\code{\link{calendar_get}},
\code{\link{calendar}()}
}
\author{
N. Frerebeau
}
\concept{time scales tools}
\keyword{internal}
