% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/era.R
\docType{methods}
\name{convert}
\alias{convert}
\alias{convert-method}
\alias{convert,character,character-method}
\alias{convert,TimeScale,TimeScale-method}
\title{Calendar Converter}
\usage{
convert(from, to, ...)

\S4method{convert}{character,character}(from, to)

\S4method{convert}{TimeScale,TimeScale}(from, to)
}
\arguments{
\item{from}{A \code{\linkS4class{TimeScale}} object describing the source era.}

\item{to}{A \code{\linkS4class{TimeScale}} object describing the target era.}

\item{...}{Currently not used.}
}
\value{
A \code{\link{function}} that when called with a single numeric argument (years)
converts years from one calendar to another.
}
\description{
Interconverts dates in a variety of calendars.
}
\note{
Adapted from \code{\link[era:yr_transform]{era::yr_transform()}} by Joe Roe.
}
\examples{
## Define time scales
BP <- era("BP")
AD <- era("AD")

## Make conversion functions
BP_to_AD <- convert(BP, AD)
AD_to_BP <- convert(AD, BP)

## Convert years
BP_to_AD(0)
AD_to_BP(1950)
}
\seealso{
Other time scales tools: 
\code{\link{era_get}},
\code{\link{era}()}
}
\author{
N. Frerebeau
}
\concept{time scales tools}
\keyword{internal}
