% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/project.R
\docType{methods}
\name{convert}
\alias{convert}
\alias{convert-method}
\alias{convert,character,character-method}
\alias{convert,TimeScale,TimeScale-method}
\title{Calendar Converter}
\usage{
convert(from, to, ...)

\S4method{convert}{character,character}(from, to, precision = getOption("chronos.precision"))

\S4method{convert}{TimeScale,TimeScale}(from, to, precision = getOption("chronos.precision"))
}
\arguments{
\item{from}{A \code{\linkS4class{TimeScale}} object describing the source calendar.}

\item{to}{A \code{\linkS4class{TimeScale}} object describing the target calendar.}

\item{...}{Currently not used.}

\item{precision}{A length-one \code{\link{integer}} vector indicating the number of
decimal places. Defaults to \code{NA}: no rounding is performed.}
}
\value{
A \code{\link{function}} that when called with a single numeric argument (years)
converts years from one calendar era to another.
}
\description{
Interconverts dates in a variety of calendars.
}
\note{
Adapted from \code{\link[era:yr_transform]{era::yr_transform()}} by Joe Roe.
}
\seealso{
Other time scales tools: 
\code{\link{calendar_get}},
\code{\link{calendar}()},
\code{\link{gregorian}},
\code{\link{julian}()}
}
\author{
N. Frerebeau
}
\concept{time scales tools}
\keyword{internal}
