% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/era.R
\docType{methods}
\name{era}
\alias{era}
\alias{era-method}
\alias{era,character-method}
\alias{era,TimeSeries-method}
\title{Time Scales}
\usage{
era(object, ...)

\S4method{era}{character}(object)

\S4method{era}{TimeSeries}(object)
}
\arguments{
\item{object}{A \code{\link{character}} string specifying the time scale (see details).}

\item{...}{Currently not used.}
}
\value{
A \code{\linkS4class{TimeScale}} object.
}
\description{
Time Scales
}
\details{
The following time scales are available:
\describe{
\item{\code{BP}}{Before Present.}
\item{\code{BC}}{Before Christ.}
\item{\code{BCE}}{Before Common Era.}
\item{\code{AD}}{Anno Domini.}
\item{\code{CE}}{Common Era.}
\item{\code{b2k}}{Years before 2000.}
}
}
\note{
Inspired by \code{\link[era:era]{era::era()}} by Joe Roe.
}
\examples{
## Define time scales
(BP <- era("BP"))
(AD <- era("AD"))
}
\seealso{
Other time scales tools: 
\code{\link{convert}()},
\code{\link{era_get}}
}
\author{
N. Frerebeau
}
\concept{time scales tools}
