% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/era.R
\docType{methods}
\name{era}
\alias{era}
\alias{era-method}
\alias{era,character-method}
\alias{era,CalibratedAges-method}
\alias{era,CalibratedSPD-method}
\title{Time Scales}
\usage{
era(object, ...)

\S4method{era}{character}(object)

\S4method{era}{CalibratedAges}(object)

\S4method{era}{CalibratedSPD}(object)
}
\arguments{
\item{object}{A \code{\link{character}} string specifying the time scale (see details).}

\item{...}{Currently not used.}
}
\value{
A \code{\linkS4class{TimeScale}} object.
}
\description{
Time Scales
}
\details{
The following time scales are available:
\describe{
\item{\code{BP}}{Before Present.}
\item{\code{BC}}{Before Christ.}
\item{\code{BCE}}{Before Common Era.}
\item{\code{AD}}{Anno Domini.}
\item{\code{CE}}{Common Era.}
\item{\code{b2k}}{Years before 2000.}
}
}
\examples{
## Define time scales
BP <- era("BP")
AD <- era("AD")

## Make conversion functions
BP_to_AD <- convert(BP, AD)
AD_to_BP <- convert(AD, BP)

## Convert years
BP_to_AD(0)
AD_to_BP(1950)
}
\seealso{
Other time scales: 
\code{\link{convert}()},
\code{\link{project}()},
\code{\link{years}()}
}
\author{
N. Frerebeau
}
\concept{time scales}
\keyword{internal}
