% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/era.R
\docType{methods}
\name{era_get}
\alias{era_get}
\alias{era_label}
\alias{era_label-method}
\alias{era_name}
\alias{era_name-method}
\alias{era_calendar}
\alias{era_calendar-method}
\alias{era_epoch}
\alias{era_epoch-method}
\alias{era_direction}
\alias{era_direction-method}
\alias{era_year}
\alias{era_year-method}
\alias{era_label,TimeScale-method}
\alias{era_label,TimeLine-method}
\alias{era_label,TimeSeries-method}
\alias{era_name,TimeScale-method}
\alias{era_name,TimeLine-method}
\alias{era_name,TimeSeries-method}
\alias{era_calendar,TimeScale-method}
\alias{era_calendar,TimeLine-method}
\alias{era_calendar,TimeSeries-method}
\alias{era_epoch,TimeScale-method}
\alias{era_epoch,TimeLine-method}
\alias{era_epoch,TimeSeries-method}
\alias{era_direction,TimeScale-method}
\alias{era_direction,TimeLine-method}
\alias{era_direction,TimeSeries-method}
\alias{era_year,TimeScale-method}
\alias{era_year,TimeLine-method}
\alias{era_year,TimeSeries-method}
\title{Era Parameters}
\usage{
era_label(object)

era_name(object)

era_calendar(object)

era_epoch(object)

era_direction(object)

era_year(object)

\S4method{era_label}{TimeScale}(object)

\S4method{era_label}{TimeLine}(object)

\S4method{era_label}{TimeSeries}(object)

\S4method{era_name}{TimeScale}(object)

\S4method{era_name}{TimeLine}(object)

\S4method{era_name}{TimeSeries}(object)

\S4method{era_calendar}{TimeScale}(object)

\S4method{era_calendar}{TimeLine}(object)

\S4method{era_calendar}{TimeSeries}(object)

\S4method{era_epoch}{TimeScale}(object)

\S4method{era_epoch}{TimeLine}(object)

\S4method{era_epoch}{TimeSeries}(object)

\S4method{era_direction}{TimeScale}(object)

\S4method{era_direction}{TimeLine}(object)

\S4method{era_direction}{TimeSeries}(object)

\S4method{era_year}{TimeScale}(object)

\S4method{era_year}{TimeLine}(object)

\S4method{era_year}{TimeSeries}(object)
}
\arguments{
\item{object}{A \code{\linkS4class{TimeScale}} object.}
}
\description{
Era Parameters
}
\examples{
## Create a custom gregorian calendar
gregorian(
  label = "AUC",
  name = "Ab urbe condita",
  epoch = 753,
  direction = 1
)
}
\seealso{
Other time scales tools: 
\code{\link{convert}()},
\code{\link{era}()},
\code{\link{gregorian}},
\code{\link{julian}()}
}
\author{
N. Frerebeau
}
\concept{time scales tools}
