% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/calendar-gregorian.R
\docType{methods}
\name{fixed}
\alias{fixed}
\alias{fixed-method}
\alias{fixed,numeric,missing,missing,GregorianCalendar-method}
\alias{fixed,numeric,numeric,numeric,GregorianCalendar-method}
\title{\emph{Rata Die} (Fixed Date)}
\usage{
fixed(year, month, day, calendar, ...)

\S4method{fixed}{numeric,missing,missing,GregorianCalendar}(year, calendar, scale = 1)

\S4method{fixed}{numeric,numeric,numeric,GregorianCalendar}(year, month, day, calendar)
}
\arguments{
\item{year}{A \code{\link{numeric}} vector of years. If \code{month} and \code{day} are missing,
decimal years are expected.}

\item{month}{A \code{\link{numeric}} vector of months.}

\item{day}{A \code{\link{numeric}} vector of days.}

\item{calendar}{A \code{\linkS4class{TimeScale}} object specifying the calendar of
\code{year}, \code{month} and \code{day} (see \code{\link[=calendar]{calendar()}}).}

\item{...}{Currently not used.}

\item{scale}{A length-one \code{\link{integer}} vector specifying the number of years
represented by one unit. It should be a power of 10 (i.e. 1000 means ka).}
}
\value{
A \code{\linkS4class{RataDie}} object.
}
\description{
\emph{Rata Die} (Fixed Date)
}
\details{
\emph{Rata die} are represented as the number of days since 01-01-01 (Gregorian),
with negative values for earlier dates.
}
\examples{
## R 1.0.0
(y <- fixed(year = 2000, month = 02, day = 29, calendar = CE()))
as_date(y, calendar = CE())
as_year(y, calendar = CE())

## Vector of years expressed in ka BP
(ka <- fixed(c(30, 35, 40), calendar = BP(), scale = 1000))
as_year(ka, calendar = CE())
}
\references{
Reingold, E. M. and Dershowitz, N. (2018). \emph{Calendrical Calculations:
The Ultimate Edition}. Cambridge University Press.
\doi{10.1017/9781107415058}.
}
\seealso{
Other fixed date tools: 
\code{\link{as_date}()},
\code{\link{as_decimal}()},
\code{\link{as_year}()},
\code{\link{fixed_gregorian}}
}
\author{
N. Frerebeau
}
\concept{fixed date tools}
