% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/calendar.R
\docType{methods}
\name{get_calendar}
\alias{get_calendar}
\alias{get_calendar-method}
\alias{set_calendar}
\alias{set_calendar-method}
\alias{get_calendar,ANY-method}
\alias{set_calendar,character-method}
\alias{set_calendar,missing-method}
\title{Get or Set the Default Calendar}
\usage{
get_calendar(...)

set_calendar(object)

\S4method{get_calendar}{ANY}(...)

\S4method{set_calendar}{character}(object)

\S4method{set_calendar}{missing}()
}
\arguments{
\item{...}{Currently not used.}

\item{object}{A \code{\link{character}} string specifying the abbreviated label of
the time scale (see \code{\link[=calendar]{calendar()}}) or an object from which to extract the
time scale.}
}
\value{
A \code{\linkS4class{TimeScale}} object.
}
\description{
Get or Set the Default Calendar
}
\examples{
## Define time scales
calendar("BP")
calendar("AD")
calendar("julian")

## Shortcuts
BP()
AD()
J()
}
\seealso{
Other calendar tools: 
\code{\link{calendar}()},
\code{\link{calendar_get}},
\code{\link{convert}()},
\code{\link{gregorian}},
\code{\link{is_calendar}()},
\code{\link{julian}()}
}
\author{
N. Frerebeau
}
\concept{calendar tools}
