% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/graph.R
\docType{methods}
\name{graph}
\alias{graph}
\alias{graph-method}
\alias{graph,TimeIntervals-method}
\title{Interval Graph}
\usage{
graph(x, ...)

\S4method{graph}{TimeIntervals}(x, aggregate = TRUE, ...)
}
\arguments{
\item{x}{A \code{\linkS4class{TimeIntervals}} object.}

\item{...}{Currently not used.}

\item{aggregate}{A \code{\link{logical}} scalar: should disjoint intervals referring to
the same event be aggregated?}
}
\value{
An \pkg{igraph} graph object.
}
\description{
Interval Graph
}
\details{
An interval graph is the graph showing intersecting intervals on a line.
As time is linear and not circular, an interval graph contains no cycles
with more than three edges and no shortcuts.
}
\note{
Experimental: might change in a future release.
}
\examples{
if (requireNamespace("igraph", quietly = TRUE)) {
  ## Seven intervals
  int <- intervals(
    start = c(1, 2, 3, 6, 9, 13, 17),
    end = c(7, 4, 15, 14, 11, 18, 19),
    calendar = CE(),
    names = c("A", "B", "C", "D", "E", "F", "G")
  )

  ## Do the intervals overlap?
  overlap(int) > 0

  ## Interval graph
  g <- graph(int)
  plot(g)
}
}
\seealso{
Other chronological reasoning tools: 
\code{\link{overlap}()},
\code{\link{relations}}
}
\author{
N. Frerebeau
}
\concept{chronological reasoning tools}
