% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/calendar-gregorian.R
\docType{methods}
\name{gregorian}
\alias{gregorian}
\alias{is_gregorian}
\alias{is_gregorian-method}
\alias{is_gregorian,TimeScale-method}
\title{Gregorian Calendar}
\usage{
is_gregorian(object)

\S4method{is_gregorian}{TimeScale}(object)
}
\arguments{
\item{object}{A \code{\linkS4class{GregorianCalendar}} object.}
}
\value{
\itemize{
\item \code{is_gregorian()} returns a \code{\link{logical}} scalar.
}
}
\description{
Gregorian Calendar
}
\examples{
## Define time scales
(BP <- calendar("BP"))
(AD <- calendar("AD"))
}
\seealso{
Other time scales tools: 
\code{\link{calendar_get}},
\code{\link{calendar}()},
\code{\link{convert}()},
\code{\link{julian}()}
}
\author{
N. Frerebeau
}
\concept{time scales tools}
