% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/calendar.R
\docType{methods}
\name{gregorian}
\alias{gregorian}
\alias{BP}
\alias{b2k}
\alias{BC}
\alias{BCE}
\alias{AD}
\alias{CE}
\title{Gregorian Calendar}
\usage{
BP(...)

b2k(...)

BC(...)

BCE(...)

AD(...)

CE(...)
}
\arguments{
\item{...}{Currently not used.}
}
\value{
A \code{\linkS4class{GregorianCalendar}} object.
}
\description{
Gregorian Calendar
}
\examples{
## Define time scales
(BP <- calendar("BP"))
(AD <- calendar("AD"))
}
\seealso{
\code{\link[=calendar]{calendar()}}

Other time scales tools: 
\code{\link{calendar_get}},
\code{\link{calendar}()},
\code{\link{convert}()},
\code{\link{julian}()}
}
\author{
N. Frerebeau
}
\concept{time scales tools}
