% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{GregorianCalendar-class}
\alias{GregorianCalendar-class}
\alias{.GregorianCalendar}
\alias{BP-class}
\alias{.BP}
\alias{b2k-class}
\alias{.b2k}
\alias{BC-class}
\alias{.BC}
\alias{BCE-class}
\alias{.BCE}
\alias{AD-class}
\alias{.AD}
\alias{CE-class}
\alias{.CE}
\title{GregorianCalendar}
\description{
An S4 class to represent a Gregorian calendar.
}
\seealso{
Other classes: 
\code{\link{Calendar-class}},
\code{\link{JulianCalendar-class}},
\code{\link{TimeLine-class}},
\code{\link{TimeSeries-class}}

Other calendar classes: 
\code{\link{Calendar-class}},
\code{\link{JulianCalendar-class}}
}
\author{
N. Frerebeau
}
\concept{calendar classes}
\concept{classes}
\keyword{internal}
