% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{GregorianCalendar-class}
\alias{GregorianCalendar-class}
\alias{.GregorianCalendar}
\title{GregorianCalendar}
\description{
An S4 class to represent a Gregorian calendar.
}
\seealso{
Other classes: 
\code{\link{JulianCalendar-class}},
\code{\link{RataDie-class}},
\code{\link{TimeIntervals-class}},
\code{\link{TimeScale-class}},
\code{\link{TimeSeries-class}}

Other calendar classes: 
\code{\link{JulianCalendar-class}},
\code{\link{TimeScale-class}}
}
\author{
N. Frerebeau
}
\concept{calendar classes}
\concept{classes}
\keyword{internal}
