% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/calendar-gregorian.R,
%   R/calendar-julian.R
\docType{methods}
\name{is}
\alias{is}
\alias{is_gregorian}
\alias{is_gregorian-method}
\alias{is_julian}
\alias{is_julian-method}
\alias{is_gregorian,TimeScale-method}
\alias{is_julian,TimeScale-method}
\title{Predicates}
\usage{
is_gregorian(object)

is_julian(object)

\S4method{is_gregorian}{TimeScale}(object)

\S4method{is_julian}{TimeScale}(object)
}
\arguments{
\item{object}{Any \R object.}
}
\value{
A \code{\link{logical}} vector.
}
\description{
Predicates
}
\seealso{
Other mutators: 
\code{\link{mutators}},
\code{\link{subset}()}
}
\author{
N. Frerebeau
}
\concept{mutators}
