% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{julian}
\alias{julian}
\title{Julian Calendar}
\arguments{
\item{...}{Currently not used.}
}
\value{
A \code{\linkS4class{JulianCalendar}} object.
}
\description{
Julian Calendar
}
\examples{
## Define time scales
(BP <- calendar("BP"))
(AD <- calendar("AD"))
}
\seealso{
\code{\link[=calendar]{calendar()}}

Other time scales tools: 
\code{\link{calendar_get}},
\code{\link{calendar}()},
\code{\link{convert}()},
\code{\link{gregorian}}
}
\author{
N. Frerebeau
}
\concept{time scales tools}
