% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/calendar-julian.R
\docType{methods}
\name{julian}
\alias{julian}
\alias{is_julian}
\alias{is_julian-method}
\alias{is_julian,TimeScale-method}
\title{Julian Calendar}
\usage{
is_julian(object)

\S4method{is_julian}{TimeScale}(object)
}
\arguments{
\item{object}{A \code{\linkS4class{JulianCalendar}} object.}
}
\value{
\itemize{
\item \code{is_julian()} returns a \code{\link{logical}} scalar.
}
}
\description{
Julian Calendar
}
\examples{
## Define time scales
(BP <- calendar("BP"))
(AD <- calendar("AD"))

## Create a custom gregorian calendar
as_gregorian(
  label = "AUC",
  name = "Ab urbe condita",
  epoch = 753,
  direction = 1
)
}
\seealso{
Other time scales tools: 
\code{\link{calendar_get}},
\code{\link{calendar}()},
\code{\link{convert}()},
\code{\link{gregorian}}
}
\author{
N. Frerebeau
}
\concept{time scales tools}
