% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/calendar.R
\docType{methods}
\name{julian}
\alias{julian}
\alias{J}
\title{Julian Calendar}
\usage{
J(...)
}
\arguments{
\item{...}{Currently not used.}
}
\value{
A \code{\linkS4class{JulianCalendar}} object.
}
\description{
Julian Calendar
}
\examples{
## Define time scales
calendar("BP")
calendar("AD")
calendar("julian")

## Shortcuts
BP()
AD()
J()
}
\seealso{
\code{\link[=calendar]{calendar()}}

Other calendar tools: 
\code{\link{calendar}()},
\code{\link{calendar_get}},
\code{\link{convert}()},
\code{\link{get_calendar}()},
\code{\link{gregorian}},
\code{\link{is}()}
}
\author{
N. Frerebeau
}
\concept{calendar tools}
