% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{JulianCalendar-class}
\alias{JulianCalendar-class}
\alias{.JulianCalendar}
\title{JulianCalendar}
\description{
An S4 class to represent a Julian calendar.
}
\seealso{
Other classes: 
\code{\link{AD-class}},
\code{\link{BC-class}},
\code{\link{BCE-class}},
\code{\link{BP-class}},
\code{\link{CE-class}},
\code{\link{GregorianCalendar-class}},
\code{\link{RataDie-class}},
\code{\link{TimeIntervals-class}},
\code{\link{TimeScale-class}},
\code{\link{TimeSeries-class}},
\code{\link{b2k-class}}

Other calendar classes: 
\code{\link{GregorianCalendar-class}},
\code{\link{TimeScale-class}}
}
\author{
N. Frerebeau
}
\concept{calendar classes}
\concept{classes}
\keyword{internal}
