% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/series.R
\docType{methods}
\name{mutators}
\alias{mutators}
\alias{get}
\alias{set}
\alias{names,TimeSeries-method}
\alias{names<-,TimeSeries-method}
\title{Get or Set Parts of an Object}
\usage{
\S4method{names}{TimeSeries}(x)

\S4method{names}{TimeSeries}(x) <- value
}
\arguments{
\item{x}{An object from which to get or set element(s).}

\item{value}{A possible value for the element(s) of \code{x}.}
}
\value{
An object of the same sort as \code{x} with the new values assigned.
}
\description{
Getters and setters to extract or replace parts of an object.
}
\seealso{
Other mutators: 
\code{\link{subset}()}
}
\author{
N. Frerebeau
}
\concept{mutators}
