% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/time.R
\docType{methods}
\name{project}
\alias{project}
\alias{project-method}
\alias{project,TimeSeries,character-method}
\alias{project,TimeSeries,TimeScale-method}
\title{Change the Time Scale}
\usage{
project(object, target, ...)

\S4method{project}{TimeSeries,character}(object, target)

\S4method{project}{TimeSeries,TimeScale}(object, target)
}
\arguments{
\item{object}{A \code{\linkS4class{TimeSeries}} object.}

\item{target}{A \code{\linkS4class{TimeSeries}} object serving as a template for the
time scale. Alternatively, a \code{\link{character}} string specifying the target
time scale or a \code{\linkS4class{TimeScale}} object (see \code{\link[=era]{era()}}).}

\item{...}{Currently not used.}
}
\value{
A \code{\linkS4class{TimeSeries}} object.
}
\description{
Change the time scale of time series.
}
\examples{
## Create time series
A <- series(rnorm(300), scale = era("CE"), start = 1000)
B <- series(rnorm(300), scale = era("CE"), start = 1000)

## Reproject to BP time scale
(A <- project(A, "BP"))
(B <- project(B, A))
}
\seealso{
Other time series tools: 
\code{\link{series}()},
\code{\link{start}()},
\code{\link{time}()},
\code{\link{window}()}
}
\author{
N. Frerebeau
}
\concept{time series tools}
