% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/project.R
\docType{methods}
\name{project}
\alias{project}
\alias{project-method}
\alias{project,TimeLine,TimeScale-method}
\alias{project,TimeLine,TimeLine-method}
\alias{project,TimeSeries,character-method}
\alias{project,TimeSeries,TimeScale-method}
\alias{project,TimeSeries,TimeLine-method}
\alias{project,TimeSeries,TimeSeries-method}
\title{Change the Time Scale}
\usage{
project(object, target, ...)

\S4method{project}{TimeLine,TimeScale}(object, target)

\S4method{project}{TimeLine,TimeLine}(object, target)

\S4method{project}{TimeSeries,TimeScale}(object, target)

\S4method{project}{TimeSeries,TimeLine}(object, target)

\S4method{project}{TimeSeries,TimeSeries}(object, target)
}
\arguments{
\item{object}{A \code{\linkS4class{TimeSeries}} object.}

\item{target}{A \code{\linkS4class{TimeSeries}} object serving as a template for the
target calendar. Alternatively, a \code{\link{character}} string specifying the target
calendar or a \code{\linkS4class{TimeScale}} object (see \code{\link[=era]{era()}}).}

\item{...}{Currently not used.}
}
\value{
A \code{\linkS4class{TimeSeries}} object.
}
\description{
Change the time scale of time series.
}
\examples{
## Create time series
A <- series(rnorm(100), era("CE"), time = 1000:1099)
B <- series(rnorm(100), era("CE"), time = 1100:1199)

## Reproject to BP time scale
(A <- project(A, era("BP")))
(B <- project(B, A))
}
\seealso{
Other time series tools: 
\code{\link{plot}()},
\code{\link{series}()},
\code{\link{start}()},
\code{\link{time}()},
\code{\link{window}()},
\code{\link{years}()}
}
\author{
N. Frerebeau
}
\concept{time series tools}
