% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/project.R
\docType{methods}
\name{project}
\alias{project}
\alias{project-method}
\alias{project,TimeSeries,character-method}
\alias{project,TimeSeries,Calendar-method}
\title{Change the Time Scale}
\usage{
project(object, target, ...)

\S4method{project}{TimeSeries,character}(object, target)

\S4method{project}{TimeSeries,Calendar}(object, target)
}
\arguments{
\item{object}{A \code{\linkS4class{TimeSeries}} object.}

\item{target}{A \code{\linkS4class{TimeSeries}} object serving as a template for the
target calendar. Alternatively, a \code{\link{character}} string specifying the target
calendar or a \code{\linkS4class{Calendar}} object (see \code{\link[=era]{era()}}).}

\item{...}{Currently not used.}
}
\value{
A \code{\linkS4class{TimeSeries}} object.
}
\description{
Change the time scale of time series.
}
\seealso{
Other time series tools: 
\code{\link{plot}()},
\code{\link{series}()},
\code{\link{start}()},
\code{\link{time}()},
\code{\link{window}()},
\code{\link{years}()}
}
\author{
N. Frerebeau
}
\concept{time series tools}
