% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhythm.R
\name{rhythm}
\alias{rhythm}
\title{Rhythm of events}
\usage{
rhythm(x, t, f, a, ...)
}
\arguments{
\item{x}{Vector of potential data points that reflect (or do not reflect) event data}

\item{t}{Vector of time points that reflect information about events}

\item{f}{Frequency of data sampling to help udnerstand the range and scale of
\code{t}, such as if its in calendar years, time stamps, or continuous time
intervals.}

\item{a}{Annotation that labels certain points as of having particular
interest, such as event status, censorship, etc.}
}
\value{
A "combined" record of the \code{rhythm} class that has validated event
and time properties.
}
\description{
The type of data that internally helps to parse data will be built and
validated on this class definition. This allos for understanding of the type
of variable (continuous or categorical) and how it is sampled (regular or
irregular intervals).
}
\details{
Both \code{x} and \code{t} must be vectors of the same length. \code{a} is presumed to be
only annotation data at certain positions. \code{f} serves as a marker of
frequency or scale of the data (e.g. if its in days, years, months, or a
smaller, more continuous scale such as seconds, minutes, etc).
}
