% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/time.R
\docType{methods}
\name{series}
\alias{series}
\alias{series-method}
\alias{series,matrix,TimeScale-method}
\alias{series,numeric,TimeScale-method}
\alias{series,data.frame,TimeScale-method}
\title{Create Time Series}
\usage{
series(object, scale, ...)

\S4method{series}{matrix,TimeScale}(object, scale, start, end = NULL, frequency = 1, names = NULL)

\S4method{series}{numeric,TimeScale}(object, scale, start, end = NULL, frequency = 1, names = NULL)

\S4method{series}{data.frame,TimeScale}(object, scale, start, end = NULL, frequency = 1, names = NULL)
}
\arguments{
\item{object}{A \code{\link{numeric}} \code{vector} or \code{matrix} of the observed time-series
values. A \code{\link{data.frame}} will be coerced to a \code{numeric} \code{matrix} via
\code{\link[=data.matrix]{data.matrix()}}.}

\item{scale}{A \code{\linkS4class{TimeScale}} object (see \code{\link[=era]{era()}}).}

\item{...}{Currently not used.}

\item{start}{A length-one \code{\link{numeric}} vector specifying the year of the
first observation.}

\item{end}{A length-one \code{\link{numeric}} vector specifying the year of the
last observation.}

\item{frequency}{A length-one \code{\link{numeric}} vector specifying the time
difference between two observations (resolution), in years.}

\item{names}{A \code{\link{character}} string specifying the names of the time
series.}
}
\value{
A \code{\linkS4class{TimeSeries}} object.
}
\description{
Create Time Series
}
\details{
\code{series()} only supports data sampled at equidistant points in time,
expressed in decimal years (1950 means 1950.0, i.e. the beginning of the
year 1950).
}
\examples{
## Univariate
z <- series(rnorm(300), era("BCE"), start = 1000)

names(z)
start(z)
end(z)
time(z)

## Multivariate
z <- series(matrix(rnorm(300), 100, 3), era("CE"), start = 1000)

names(z)
start(z)
end(z)
time(z)
}
\seealso{
Other time series tools: 
\code{\link{project}()},
\code{\link{start}()},
\code{\link{time}()},
\code{\link{window}()}
}
\author{
N. Frerebeau
}
\concept{time series tools}
