% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/time.R
\docType{methods}
\name{span}
\alias{span}
\alias{span-method}
\alias{span,TimeSeries-method}
\title{Duration of Time Series}
\usage{
span(x)

\S4method{span}{TimeSeries}(x)
}
\arguments{
\item{x}{A \code{\linkS4class{TimeSeries}} object.}
}
\value{
A \code{\link{numeric}} vector.
}
\description{
Get the duration expressed in \emph{rata die}.
}
\examples{
## Create time-series of 20 observations

## Univariate
## Sampled every years starting from 1029 BCE
(X <- series(rnorm(30), time = 1029:1000, calendar = BCE()))

start(X)
end(X)
span(X)
time(X)
years(X) # Default calendar (Gregorian Common Era)

## Multivariate
## Sampled every century starting from 1000 CE
(Y <- series(matrix(rnorm(90), 30, 3), time = 1000:1029, calendar = CE()))

start(Y)
end(Y)
span(Y)
time(Y)
years(Y) # Default calendar (Gregorian Common Era)
}
\seealso{
Other time series tools: 
\code{\link{plot}()},
\code{\link{series}()},
\code{\link{start}()},
\code{\link{time}()},
\code{\link{window}()}
}
\author{
N. Frerebeau
}
\concept{time series tools}
