% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/time.R
\docType{methods}
\name{start}
\alias{start}
\alias{start-method}
\alias{end-method}
\alias{start,TimeLine-method}
\alias{start,TimeSeries-method}
\alias{end,TimeLine-method}
\alias{end,TimeSeries-method}
\title{Terminal Times of Time Series}
\usage{
\S4method{start}{TimeLine}(x)

\S4method{start}{TimeSeries}(x)

\S4method{end}{TimeLine}(x)

\S4method{end}{TimeSeries}(x)
}
\arguments{
\item{x}{A \code{\linkS4class{TimeSeries}} object.}
}
\value{
A \code{\link{numeric}} vector.
}
\description{
Get the times the first and last observations were taken.
}
\examples{
## Create time-series of 100 observations

## Univariate
## Sampled every years starting from 1000 CE
(X <- series(rnorm(100), era("BCE"), time = 1000:1099))

start(X)
end(X)
time(X)

## Multivariate
## Sampled every century starting from 1000 CE
## Expressed in kilo years
(Y <- series(matrix(rnorm(300), 100, 3), era("CE"), time = 1000:1099))

start(Y)
end(Y)
time(Y)
}
\seealso{
Other time series tools: 
\code{\link{plot}()},
\code{\link{project}()},
\code{\link{series}()},
\code{\link{time}()},
\code{\link{window}()},
\code{\link{years}()}
}
\author{
N. Frerebeau
}
\concept{time series tools}
