% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/time.R
\docType{methods}
\name{start}
\alias{start}
\alias{start-method}
\alias{end-method}
\alias{start,RataDie-method}
\alias{start,TimeSeries-method}
\alias{end,RataDie-method}
\alias{end,TimeSeries-method}
\title{Terminal Times of Time Series}
\usage{
\S4method{start}{RataDie}(x)

\S4method{start}{TimeSeries}(x)

\S4method{end}{RataDie}(x)

\S4method{end}{TimeSeries}(x)
}
\arguments{
\item{x}{A \code{\linkS4class{TimeSeries}} object.}
}
\value{
A \code{\link{numeric}} vector.
}
\description{
Get the times (expressed in \emph{rata die}) the first and last observations were
taken.
}
\examples{
## Create time-series of 100 observations

## Univariate
## Sampled every years starting from 1099 BCE
(X <- series(rnorm(100), time = 1099:1000, calendar = BCE()))

start(X)
end(X)
time(X)

## Multivariate
## Sampled every century starting from 1000 CE
(Y <- series(matrix(rnorm(300), 100, 3), time = 1000:1099, calendar = CE()))

start(Y)
end(Y)
time(Y)
}
\seealso{
Other time series tools: 
\code{\link{plot}()},
\code{\link{series}()},
\code{\link{time}()},
\code{\link{window}()}
}
\author{
N. Frerebeau
}
\concept{time series tools}
