% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/time.R
\docType{methods}
\name{start}
\alias{start}
\alias{start-method}
\alias{end-method}
\alias{start,TimeSeries-method}
\alias{end,TimeSeries-method}
\title{Terminal Times of Time Series}
\usage{
\S4method{start}{TimeSeries}(x)

\S4method{end}{TimeSeries}(x)
}
\arguments{
\item{x}{A \code{\linkS4class{TimeSeries}} object.}
}
\value{
A \code{\link{numeric}} vector.
}
\description{
Get the times the first and last observations were taken.
}
\seealso{
Other time series tools: 
\code{\link{project}()},
\code{\link{series}()},
\code{\link{time}()},
\code{\link{window}()}
}
\author{
N. Frerebeau
}
\concept{time series tools}
