% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/time.R
\docType{methods}
\name{start}
\alias{start}
\alias{start-method}
\alias{end-method}
\alias{start,TimeSeries-method}
\alias{end,TimeSeries-method}
\title{Terminal Times of Time Series}
\usage{
\S4method{start}{TimeSeries}(x, calendar = NULL)

\S4method{end}{TimeSeries}(x, calendar = NULL)
}
\arguments{
\item{x}{A \code{\linkS4class{TimeSeries}} object.}

\item{calendar}{A \code{\linkS4class{TimeScale}} object specifying the target calendar
(see \code{\link[=calendar]{calendar()}}). If \code{NULL} (the default), \emph{rata die} are returned.}
}
\value{
A \code{\link{numeric}} vector.
}
\description{
Get the times the first and last observations were taken.
}
\examples{
## Create time-series of 20 observations

## Univariate
## Sampled every years starting from 1029 BCE
(X <- series(rnorm(30), time = 1029:1000, calendar = BCE()))

start(X)
end(X)
time(X)

## Multivariate
## Sampled every century starting from 1000 CE
(Y <- series(matrix(rnorm(90), 30, 3), time = 1000:1029, calendar = CE()))

start(Y, calendar = CE())
end(Y, calendar = CE())
span(Y, calendar = CE())
time(Y, calendar = CE())

## Coerce to data frame
df <- as.data.frame(Y, calendar = b2k())
head(df)
}
\seealso{
Other time series tools: 
\code{\link{axis_year}()},
\code{\link{data.frame}},
\code{\link{plot}()},
\code{\link{series}()},
\code{\link{span}()},
\code{\link{time}()},
\code{\link{window}()}
}
\author{
N. Frerebeau
}
\concept{time series tools}
