% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/series.R
\docType{methods}
\name{time}
\alias{time}
\alias{time-method}
\alias{time,TimeSeries-method}
\alias{frequency,TimeSeries-method}
\title{Sampling Times of Time Series}
\usage{
\S4method{time}{TimeSeries}(x)

\S4method{frequency}{TimeSeries}(x)
}
\arguments{
\item{x}{A \code{\linkS4class{TimeSeries}} object.}
}
\value{
A \code{\link{numeric}} vector.
}
\description{
\itemize{
\item \code{time()} creates the vector of times at which a time series was sampled.
\item \code{frequency()} returns the number of samples per unit time.
}
}
\examples{
## Create time-series of 100 observations
## Sampled every years starting from 1000 CE

## Univariate
X <- series(rnorm(300), era("BCE"), start = 1000)

names(X)
start(X)
end(X)
time(X)

## Multivariate
Y <- series(matrix(rnorm(300), 100, 3), era("CE"), start = 1000)

names(Y)
start(Y)
end(Y)
time(Y)
}
\seealso{
Other time series tools: 
\code{\link{plot}()},
\code{\link{project}()},
\code{\link{series}()},
\code{\link{start}()},
\code{\link{window}()}
}
\author{
N. Frerebeau
}
\concept{time series tools}
