% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/time.R
\docType{methods}
\name{time}
\alias{time}
\alias{time-method}
\alias{time,TimeLine-method}
\alias{time,TimeSeries-method}
\alias{frequency,TimeSeries-method}
\alias{frequency,TimeLine-method}
\title{Sampling Times of Time Series}
\usage{
\S4method{time}{TimeLine}(x)

\S4method{time}{TimeSeries}(x)

\S4method{frequency}{TimeSeries}(x)

\S4method{frequency}{TimeSeries}(x)
}
\arguments{
\item{x}{A \code{\linkS4class{TimeSeries}} object.}
}
\value{
A \code{\link{numeric}} vector.
}
\description{
\itemize{
\item \code{time()} creates the vector of times at which a time series was sampled.
\item \code{frequency()} returns the number of samples per unit time.
}
}
\examples{
## Create time-series of 100 observations

## Univariate
## Sampled every years starting from 1000 CE
(X <- series(rnorm(100), calendar("BCE"), time = 1000:1099))

start(X)
end(X)
time(X)

## Multivariate
## Sampled every century starting from 1000 CE
## Expressed in kilo years
(Y <- series(matrix(rnorm(300), 100, 3), calendar("CE"), time = 1000:1099))

start(Y)
end(Y)
time(Y)
}
\seealso{
Other time series tools: 
\code{\link{as_date}()},
\code{\link{as_fixed}()},
\code{\link{as_year}()},
\code{\link{plot}()},
\code{\link{project}()},
\code{\link{series}()},
\code{\link{start}()},
\code{\link{window}()},
\code{\link{years}()}
}
\author{
N. Frerebeau
}
\concept{time series tools}
