% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/time.R
\docType{methods}
\name{time}
\alias{time}
\alias{time-method}
\alias{frequency-method}
\alias{time,TimeSeries-method}
\alias{frequency,TimeSeries-method}
\title{Sampling Times of Time Series}
\usage{
\S4method{time}{TimeSeries}(x, calendar = NULL)

\S4method{frequency}{TimeSeries}(x)
}
\arguments{
\item{x}{A \code{\linkS4class{TimeSeries}} object.}

\item{calendar}{A \code{\linkS4class{TimeScale}} object specifying the target calendar
(see \code{\link[=calendar]{calendar()}}). If \code{NULL} (the default), \emph{rata die} are returned.}
}
\value{
A \code{\link{numeric}} vector.
}
\description{
Get the sampling times:
\itemize{
\item \code{time()} creates the vector of times at which a time series was sampled
(expressed in \emph{rata die}).
\item \code{frequency()} returns the mean number of samples per unit time.
}
}
\examples{
## Create time-series of 20 observations

## Univariate
## Sampled every years starting from 1029 BCE
(X <- series(rnorm(30), time = 1029:1000, calendar = BCE()))

start(X)
end(X)
time(X)

## Multivariate
## Sampled every century starting from 1000 CE
(Y <- series(matrix(rnorm(90), 30, 3), time = 1000:1029, calendar = CE()))

start(Y, calendar = CE())
end(Y, calendar = CE())
span(Y, calendar = CE())
time(Y, calendar = CE())
}
\seealso{
Other time series tools: 
\code{\link{axis_year}()},
\code{\link{plot}()},
\code{\link{series}()},
\code{\link{span}()},
\code{\link{start}()},
\code{\link{window}()}
}
\author{
N. Frerebeau
}
\concept{time series tools}
