% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{TimeSeries-class}
\alias{TimeSeries-class}
\alias{.TimeSeries}
\title{TimeSeries}
\description{
An S4 class to represent time series.
}
\details{
It is a matrix that represents data sampled at equidistant points in time,
according to a given time scale.
}
\section{Slots}{

\describe{
\item{\code{time_labels}}{A \code{\link{character}} string specifying the names of the time
series.}

\item{\code{time_start}}{A \code{\link{numeric}} value specifying the year of the
first observation.}

\item{\code{time_increment}}{A \code{\link{numeric}} value specifying the time difference
between two observations (resolution), in years.}
}}

\note{
This class inherits from \code{\link{matrix}} and \code{\linkS4class{TimeScale}}.
}
\seealso{
Other classes: 
\code{\link{CalibratedAges-class}},
\code{\link{CalibratedSPD-class}},
\code{\link{TimeScale-class}}
}
\author{
N. Frerebeau
}
\concept{classes}
\keyword{internal}
