% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{TimeSeries-class}
\alias{TimeSeries-class}
\alias{.TimeSeries}
\title{TimeSeries}
\description{
An S4 class to represent time series.
}
\details{
It is a matrix that represents data sampled at equidistant points in time,
according to a given time scale.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{A \code{numeric} \code{\link{matrix}} giving the observed time-series values.}

\item{\code{time}}{A \code{\linkS4class{TimeLine}} object.}
}}

\note{
This class inherits from \code{\link{matrix}}.
}
\seealso{
Other classes: 
\code{\link{Calendar-class}},
\code{\link{GregorianCalendar-class}},
\code{\link{JulianCalendar-class}},
\code{\link{TimeLine-class}}

Other time classes: 
\code{\link{TimeLine-class}}
}
\author{
N. Frerebeau
}
\concept{classes}
\concept{time classes}
\keyword{internal}
