% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{TimeSeries-class}
\alias{TimeSeries-class}
\alias{.TimeSeries}
\title{TimeSeries}
\description{
An S4 class to represent time series.
}
\details{
It is a matrix that represents data sampled at equidistant points in time,
according to a given time scale.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{A \code{numeric} \code{\link{matrix}} giving the the observed time-series
values.}

\item{\code{era_label}}{A \code{\link{character}} string specifying the abbreviated label of
the time scale.}

\item{\code{era_name}}{A \code{\link{character}} string specifying the name of the time scale.}

\item{\code{era_epoch}}{A \code{\link{numeric}} value specifying the epoch year from which
years are counted (in Gregorian astronomical years).}

\item{\code{era_scale}}{An \code{\link{integer}} specifying the number of years represented by
one unit. It should be a power of 10 (i.e. 1000 means ka).}

\item{\code{era_direction}}{An \code{\link{integer}} specifying if years are counted backwards
(\code{-1}) or forwards (\code{1}) from \code{epoch}.}

\item{\code{era_unit}}{A \code{\link{character}} string specifying the name of the year unit.}

\item{\code{era_days}}{A \code{\link{numeric}} vector giving the average length of the year in
solar days.}

\item{\code{time_labels}}{A \code{\link{character}} string specifying the names of the time
series.}

\item{\code{time_start}}{A \code{\link{numeric}} value specifying the year of the
first observation.}

\item{\code{time_increment}}{A \code{\link{numeric}} value specifying the time difference
between two observations (resolution), in years.}
}}

\note{
This class inherits from \code{\link{matrix}} and \code{\linkS4class{TimeScale}}.
}
\seealso{
Other classes: 
\code{\link{TimeScale-class}}
}
\author{
N. Frerebeau
}
\concept{classes}
\keyword{internal}
