% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/time.R
\docType{methods}
\name{window}
\alias{window}
\alias{window-method}
\alias{window,TimeSeries-method}
\title{Time Series Windows}
\usage{
\S4method{window}{TimeSeries}(x, start = NULL, end = NULL)
}
\arguments{
\item{start}{A length-one \code{\link{numeric}} vector specifying the start time of the
period of interest.}

\item{end}{A length-one \code{\link{numeric}} vector specifying the end time of the
period of interest.}
}
\value{
A \code{\linkS4class{TimeSeries}} object.
}
\description{
Extracts the subset of the object \code{x} observed between the times \code{start} and
\code{end}.
}
\examples{
## Create time series
mtx <- matrix(rnorm(300), 100, 3)
(x <- series(mtx, era("CE"), start = 1000, increment = 1))

## Subset
(y <- window(x, start = 1025, end = 1050))
}
\seealso{
Other time series tools: 
\code{\link{project}()},
\code{\link{series}()},
\code{\link{start}()},
\code{\link{time}()}
}
\author{
N. Frerebeau
}
\concept{time series tools}
