% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/years.R
\docType{methods}
\name{years}
\alias{years}
\alias{years-method}
\alias{years,numeric,TimeScale-method}
\alias{years,TimeSeries,missing-method}
\title{Year Vectors}
\usage{
years(object, calendar, ...)

\S4method{years}{numeric,TimeScale}(object, calendar, scale = 1, sort = FALSE)

\S4method{years}{TimeSeries,missing}(object)
}
\arguments{
\item{object}{A \code{\link{numeric}} vector of (decimal) years.}

\item{calendar}{A \code{\linkS4class{TimeScale}} object (see \code{\link[=era]{era()}}).}

\item{...}{Currently not used.}

\item{scale}{A length-one \code{\link{integer}} vector specifying the number of years
represented by one unit. It should be a power of 10 (i.e. 1000 means ka).}

\item{sort}{A \code{\link{logical}} scalar: should data be sorted in chronological
order?}
}
\value{
A \code{\linkS4class{TimeLine}} object.
}
\description{
Year Vectors
}
\details{
\code{years()} only supports data sampled at equidistant points in time,
expressed in decimal years (1950 means 1950.0, i.e. the beginning of the
year 1950).
}
\examples{
## R foundation
(y <- years(c(1994, 2000), era("CE")))
format(y)

## Vector of years expressed in ka BP
(ka <- years(c(30, 35, 40), era("BP"), scale = 1000))
format(ka)
format(ka, format = "ka")
}
\seealso{
Other time series tools: 
\code{\link{plot}()},
\code{\link{project}()},
\code{\link{series}()},
\code{\link{start}()},
\code{\link{time}()},
\code{\link{window}()}
}
\author{
N. Frerebeau
}
\concept{time series tools}
