% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/years.R
\docType{methods}
\name{years}
\alias{years}
\alias{years-method}
\alias{years,numeric,Calendar-method}
\title{Year Vectors}
\usage{
years(object, calendar, ...)

\S4method{years}{numeric,Calendar}(object, calendar, scale = 1, sort = FALSE)
}
\arguments{
\item{object}{A \code{\link{numeric}} vector of (decimal) years.}

\item{calendar}{A \code{\linkS4class{Calendar}} object (see \code{\link[=calendar]{calendar()}}).}

\item{...}{Currently not used.}

\item{scale}{A length-one \code{\link{integer}} vector specifying the number of years
represented by one unit. It should be a power of 10 (i.e. 1000 means ka).
Will be coerced with \code{\link[=as.integer]{as.integer()}} (hence truncated toward zero).}

\item{sort}{A \code{\link{logical}} scalar: should data be sorted in chronological
order?}
}
\value{
A \code{\linkS4class{TimeLine}} object.
}
\description{
Year Vectors
}
\details{
\code{years()} only supports data sampled at equidistant points in time,
expressed in decimal years (1950 means 1950.0, i.e. the beginning of the
year 1950).
}
\examples{
## R foundation
(y <- years(1994, calendar("CE")))
}
\seealso{
Other time series tools: 
\code{\link{plot}()},
\code{\link{project}()},
\code{\link{series}()},
\code{\link{start}()},
\code{\link{time}()},
\code{\link{window}()}
}
\author{
N. Frerebeau
}
\concept{time series tools}
